CombinedParsers.jl Documentation

A package for combining parsers and transforming strings into julia types.

Compose parsers with the functional parser combinator paradigm, utilize Julia's type inferrence for transformations, log conveniently for debugging, and let Julia compile your parser for good performance.

Note

CombinedParsers.jl is currently an α release. The first official released is prepared for JuliaCon2020.

Package Features

Clear syntax integrates grammar and transformations with Julia type inference.

Higher-order parsers depending on the parsing state allow for not context-free parsers.

All valid parsings can be iterated lazily.

Interoperable with TextParse.jl: existing TextParse.AbstractToken implementations can be used with CombinedParsers. CombinedParsers.AbstractParser provide TextParse.tryparsenext and can be used e.g. in CSV.jl.

Parametric parser and state types enable Julia compiler optimizations.

Compiled regular expression parsers in pure julia are provided with the re_str macro. CombinedParsers.jl is tested against the PCRE C library testset.

AbstractTrees.jl interface provides colored and clearly layed out printing in the REPL.

Convenient logging of the parsing process with NamedParsers and SideeffectParsers.

CombinedParsers generalize from strings to parsing any type supporting getindex, nextind, prevind methods.

Writing Parsers

CombinedParsers provides constructors to combine parsers and transform (sub-)parsings arbitrarily with julia syntax. Combinator constructors are discussed in the user manual.

julia> # Julia eval for +-*/ operators
       function eval_ops((l,opr))
           for (op,val) in opr
               l = eval( Expr(:call, Symbol(op), l, val))
           end
           l::Rational{Int}
       end
eval_ops (generic function with 1 method)

julia> using CombinedParsers

julia> using TextParse
ERROR: ArgumentError: Package TextParse not found in current path:
- Run `import Pkg; Pkg.add("TextParse")` to install the TextParse package.

julia> @with_names begin
           number = map(Rational{Int}, TextParse.Numeric(Int))
           factor = Either(number)  # or expression in parenthesis, see push! below
           divMul = map(eval_ops,
                        Sequence( factor, Repeat( CharIn("*/"), factor ) ) )
           addSub = map(eval_ops,
       		 divMul * Repeat( CharIn("+-") * divMul ) )
           parens = Sequence(2, "(",addSub,")" )
           push!(factor, parens)
           expr = (addSub * AtEnd())[1]
       end;
ERROR: UndefVarError: TextParse not defined

julia> parse(log_names(expr), "1/((1+2)*4+3*(5*2))")
ERROR: UndefVarError: expr not defined

Is every rational answer ultimately the inverse of a universal question in life?

Package Guide

The Overview provides a tutorial explaining how to get started using CombinedParsers.

Some examples of packages using Documenter can be found on the Examples page.

See the Index for the complete list of documented functions and types.

Getting Started

Basics

Character Sets

Sequence

Either

Repeat

Assertions

Atomic groups

Parsing Regular Expressions and constructing an equivalent AbstractParser

Repeatable patterns, Optional, Lazy and Atomic

Sequences and Alternations

Captures

LookAround

Library Outline

Public Documentation

Public Interface

CombinedParsers

CombinedParsers.Regexp

Parsing, Logging and Side-Effects

Transformations

Parser Constructors

Character Matchers

Combining Parser

Repeating

Assertions

Printing

Internals

Parser Types

Abstract Types

Rewriting Parsers

Regular Expression Types

Iterating Parsings

Acknowledgements

The work was inspired by Scala fastparse package and the Julia parsing packages

Parsers.jl

A collection of parsers for date and primitive types.

TextParse.jl

A bunch of fast text parsing tools.

used in CSV.jl

Nullables.jl

TextParse.jl integrates with CombinedParsers.jl both ways by type CombinedParsers.AbstractParser <: TextParse.AbstractToken and providing a method for TextParse.tryparsenext.

Dates.tryparsenext — Function

TextParse.tryparsenext(x::AbstractParser,str,i,till,opts=TextParse.default_opts)

TextParse.jl integrates with CombinedParsers.jl both ways.

julia> p = ("Number:" * Repeat(' ') * TextParse.Numeric(Int))[3]
🗄 Sequence |> map(#37)
├─ Number\:
├─ \ *  |> Repeat
└─ <Int64>
::Int64

julia> parse(p, "Number:    42")
42

julia> TextParse.tryparsenext(p, "Number:    42")
(Nullable{Int64}(42), 14)

Automa.jl

A Julia package for text validation, parsing, and tokenizing based on state machine compiler. The package compiles a parser deterministic finite automaton. Processing of parsed tokens is done with actions. No UTF8 support.

ParserCombinator.jl

old source base (pre 2016, fixed for Julia 1.0 in 2018)

using Nullables

no iterator API

performance

mutable matcher types

matcher types not parametric

Index

CombinedParsers

CombinedParsers.Regexp

CombinedParsers.Always

CombinedParsers.AnyChar

CombinedParsers.Atomic

CombinedParsers.CharIn

CombinedParsers.CharNotIn

CombinedParsers.Either

CombinedParsers.JoinSubstring

CombinedParsers.MemoTreeChildren

CombinedParsers.NegativeLookahead

CombinedParsers.NegativeLookbehind

CombinedParsers.Never

CombinedParsers.Optional

CombinedParsers.PositiveLookahead

CombinedParsers.PositiveLookbehind

CombinedParsers.Repeat

CombinedParsers.Sequence

Base.:!

Base.:|

Base.:|

Base.convert

Base.join

Base.map

Base.match

Base.parse

Base.tryparse

CombinedParsers.Repeat1

CombinedParsers.Repeat_stop

CombinedParsers.Repeat_until

CombinedParsers.alternate

CombinedParsers.log_names

CombinedParsers.map_at

CombinedParsers.parser

CombinedParsers.print_constructor

CombinedParsers.regex_escape

CombinedParsers.regex_inner

CombinedParsers.regex_prefix

CombinedParsers.regex_string

CombinedParsers.regex_suffix

CombinedParsers.sEither

CombinedParsers.sSequence

CombinedParsers.with_effect

CombinedParsers.with_log

CombinedParsers.with_name

CombinedParsers.@with_names

CombinedParsers.Regexp.@re_str