Iterating Parsings

CombinedParsers._iterate — Function

_iterate(parser, sequence, till, i, states)

Note: i is the index in sequence after parser match according to state (and not the start of the match), such that start_index(sequence,after,parser,state) returns the start of the matching subsequence, and sequence[start_index(sequence,after,parser,state):prevind(sequence,i)] is the matched subsequence.

_iterate(p::ParserWithCaptures, sequence::AbstractString)

Missing docstring.

Missing docstring for CombinedParsers.get. Check Documenter's build log for details.