Parser Types

Abstract Types

CombinedParsers.ParserTypes — Constant

Julia types that can convert(AbstractToken,x). TODO: remove

CombinedParsers.AbstractParser — Type

AbstractParser{T} <: AbstractToken{T}

Abstract parser type for parsers returning matches transformed to ::T.

CombinedParsers.LeafParser — Type

LeafParser{T} <: AbstractToken{T}

Abstract parser type for parsers that have no sub-parser. Used for dispatch in deepmap_parser

CombinedParsers.ConstantParser — Type

wrapper for stepping with ncodeunit length.

CombinedParsers.NIndexParser — Type

Abstract type for stepping with previndex/nextindex, accounting for ncodeunit length of chars at point.

CombinedParsers.WrappedParser — Type

Abstract type for parser wrappers, providing default methods

CombinedParsers.LookAround — Type

Parsers that do not consume any input can inherit this type.

CombinedParsers.Transformation — Type

Transformation{T}(transform::Function, p_) where {T}
map_at(f::Function, p, a...)
map_at(f::Function, Tc::Type, p, a...)
instance(Tc::Type, p::ParserTypes, a...)
instance(Tc::Type, p::ParserTypes)
Base.map(f::Function, Tc::Type, p::ParserTypes, a...)
Base.map(f::Function, p::ParserTypes, a...)

Parser transforming result of a wrapped parser. a... is passed as additional arguments to f (at front .

Rewriting Parsers

CombinedParsers.deepmap_parser — Function

deepmap_parser(f::Function,mem::AbstractDict,x::LeafParser,a...)

return

    get!(mem,x) do
        f(x,a...)
    end

deepmap_parser(f::Function,mem::AbstractDict,x,a...)

Perform a deep transformation of a CombinedParser.

Note

For a custom parser P<:AbstractParser with sub-parsers, provide a method

deepmap_parser(f::Function,mem::AbstractDict,x::P,a...) =
    get!(mem,x) do
        ## construct replacement, e.g. if P <: WrappedParser
        P(deepmap_parser(f,mem,x.parser,a...))
    end

deepmap_parser(f::Function,x::AbstractParser,a...)

Perform a deep transformation of a CombinedParser. Used for log_names.

Calls deepmap_parser(f,IdDict(),x,a...).

Regular Expression Types

CombinedParsers.Regexp.ParseMatch — Type

Wrapper type for SequenceWithCaptures, providing getindex and getproperty behavior like RegexMatch.

julia> m = match(re"(?<a>so)+ (or)", "soso or")
ParseMatch("soso or", a="so", 2="or")

julia> m[:a]
"so"

julia> m[2]
"or"

julia> m.match, m.captures
("soso or", SubString{String}["so", "or"])

CombinedParsers.Regexp.Capture — Type

Capture a parser result, optionally with a name. index field is initialized when calling ParserWithCaptures on the parser.

ParserWithCaptures

CombinedParsers.Regexp.Backreference — Type

Backreference(f::Function,index::Integer)

Backreference(f::Function,name::Union{Nothing,Symbol},index::Integer)

Backreference(f::Function,name::AbstractString)

Parser matching previously captured sequence, optionally with a name. index field is recursively set when calling 'ParserWithCaptures` on the parser.

CombinedParsers.Regexp.Subroutine — Type

Parser matching preceding capture, optionally with a name. index field is recursively set when calling 'ParserWithCaptures` on the parser.

CombinedParsers.Regexp.Conditional — Type

Conditional parser, _iterate cycles conditionally on _iterate_condition through matches in field yes and no respectively.

CombinedParsers.Regexp.DupSubpatternNumbers — Type

Parser wrapper for ParserWithCaptures, setting resetindex=true in `deepmapparser(::typeof(indexedcaptures),...)`.

CombinedParsers.Regexp.ParserWithCaptures — Type

Top level parser supporting regular expression features captures, backreferences and subroutines. Collects subroutines in field subroutines::Vector and indices of named capture groups in field names::Dict.

Note

implicitly called in match

See also Backreference, Capture, Subroutine

CombinedParsers.Regexp.SequenceWithCaptures — Type

SequenceWithCaptures ensapsulates a sequence to be parsed, and parsed captures.

See also ParserWithCaptures