Public Documentation

Documentation for CombinedParsers.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Matching

CombinedParsers — Module

A package for combining parsers and transforming strings into julia types.

Compose parsers with the functional parser combinator paradigm, utilize Julia's type inferrence for transformations, log conveniently for debugging, and let Julia compile your parser for good performance.

Base.match — Function

Base.match(parser::ParserTypes,sequence::AbstractString[, idx::Integer]; log=nothing)

Plug-in replacement for match(::Regex,sequence).

If log!==nothing, parser is transformed with log_names(p, log). See also log_names.

CombinedParsers.ParseMatch — Type

Wrapper type for CombinedParsers.Regexp.SequenceWithCaptures, providing getindex and getproperty behavior like RegexMatch.

julia> m = match(re"(?<a>so)+ (or)", "soso or")
ParseMatch("soso or", a="so", 2="or")

julia> m[:a]
"so"

julia> m[2]
"or"

julia> m.match, m.captures
("soso or", SubString{String}["so", "or"])

Base.iterate — Function

iterate(m::CombinedParser,s::AbstractString)

Returns iterate(ParseMatch(m,s,1,1,nothing)).

See also ParseMatch.

Base.iterate(x::ParseMatch[, m::ParseMatch=x])

Returns next match at m.start after m.state, see _iterate(m).

See also ParseMatch.

Base.iterate(x::MatchesIterator[, s::ParseMatch=ParseMatch(x,x.from,x.from,nothing)])

Iterate match s at current position. While no match is found and s.start<=x.till, increase s.start. Return first next ParseMatch (as return value and state) or nothing when at m.till.

CombinedParsers.match_all — Function

match_all(parser::ParserTypes, sequence, idx=1; log=nothing)

Returns an iterator over all matches at all indices in the sequence.

Parsing

Base.get — Function

Base.get(x::ParseMatch{<:MatchTuple})

Get the result of a match result.

julia> m = match(re"(?<a>so)+ (or)", "soso or")
ParseMatch("soso or", a="so", 2="or")

julia> get(m)
([('s', 'o'), ('s', 'o')], ' ', ('o', 'r'))

julia> m[2]
"or"

julia> m.match, m.captures
("soso or", SubString{String}["so", "or"])

Base.get(parser::Bytes{T}, sequence::Vector{UInt8})

Note

Endianness can be achieved by just mapping bswap

julia> map(bswap, Bytes(2,UInt16))([0x16,0x11])
0x1611

julia> Bytes(2,UInt16)([0x16,0x11])
0x1116

Base.get(parser::AbstractToken{Nothing}, sequence, till, after, i, state)

Default method for parser types returning nothing

Base.get(parser::Transformation{MatchRange}, a...)

Does not evaluate get(parser.transform,...).

Base.get(parser::Transformation{<:Constant}, a...)

does not evaluate get(parser.transform,...).

Base.get(parser::Transformation{<:Function}, a...)
Base.get(parser::Transformation{<:Type}, a...)

Function call parser.transform(get(parser.parser,a...)).

Base.get(parser::Transformation{<:IndexAt}, a...)

getindex(get(parser.parser,a...).parser.transform)

Base.parse — Function

parse(parser::ParserTypes, str::AbstractString; log=nothing)

Parse a string with a CombinedParser as an instance of result_type(parser).

If log is a Vector{Symbol}, parser is transformed with log_names(p, log). See also log_names.

DocTestFilters = r"map\(.+\)"

julia> using TextParse

julia> p = ("Number: "*TextParse.Numeric(Int))[2]
🗄 Sequence[2]
├─ Number\:\
└─ <Int64>
::Int64

julia> parse(p,"Number: 42")
42

Base.tryparse — Function

tryparse(parser::ParserTypes, str::AbstractString[, idx=1])

Like parse, but returns either a value of result_type(parser) or nothing if string does not start with with a match.

CombinedParsers.tryparse_pos — Function

tryparse_pos(parser::ParserTypes, str::AbstractString[, idx=1, till=lastindex(s)])

Like parse, but returns either a tuple of result_type(parser) and the position after the match, or nothing if string does not start with with a match.

CombinedParsers.parse_all — Function

parse_all(parser::ParserTypes, sequence, idx=1)

Returns an iterator over all parsings of the sequence starting at idx.

Typed Transformation Parsers

Base.:! — Function

(!)(x::AbstractToken)

Parser Transformation getting the matched SubString.

julia> parse(Repeat(CharIn(:L)),"abc123")
3-element Array{Char,1}:
 'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)
 'b': ASCII/Unicode U+0062 (category Ll: Letter, lowercase)
 'c': ASCII/Unicode U+0063 (category Ll: Letter, lowercase)

julia> parse(!Repeat(CharIn(:L)),"abc123")
"abc"

(!)(x::JoinSubstring)

Parser transformating result v -> InternedStrings.intern(v).

(!)(x::NamedParser)

Parser transformating result v -> v=>x.name.

CombinedParsers.JoinSubstring — Type

JoinSubstring(x)
(!)(x::AbstractToken)

Parser Transformation getting the matched SubString.

Base.map — Function

map(f::Function, p::AbstractToken, a...)

Parser matching p, transforming parsing results (x) with function f(x,a...).

See also: get, Transformation

map(T::Type, p::AbstractToken, a...)

Parser matching p, transforming ps parsing result with constructor T(x,a...).

See also: get, Transformation

map(index::IndexAt, p::AbstractToken, a...)
map(constant, p::AbstractToken, a...)

Parser matching p, transforming ps parsing results to getindex(x,index) or constant.

See also: get, Transformation

CombinedParsers.MatchRange — Type

MatchRange(p::AbstractToken)

Succeed iif p succeeds, if so results in sequence match index UnitRange.

CombinedParsers.Constant — Type

A Transformation{<:Constant} skips evaluation of get(.parser).

CombinedParsers.IndexAt — Type

Struct for fast access to an index of a Transformation.

julia> using CombinedParsers.Regexp

julia> p = re"(?:a|b*)."[1]
🗄 Sequence[1]
├─ |🗄... Either
│  ├─ a 
│  └─ b*  |> Repeat
└─ [^\n] CharNotIn
::Union{Char, Array{Char,1}}

See also getindex, Sequence.

CombinedParsers.result_type — Function

result_type(::Type{T}) where {T<:Tuple}

Internally used for Sequence result_type.

CombinedParsers.infer_result_type — Function

infer_result_type(f::Function,Tc::Type,p::ParserTypes,onerror::AbstractString,ts::Type...; throw_empty_union=true)

Used by Parser Transformations to infer result type of a parser. Throws error if type inference fails, if throwemptyunion=true.

CombinedParsers.state_type — Function

CombinedParsers.state_type(x::T)

Return the state type of x

Parser Constructors

CombinedParsers.parser — Function

parser(x::Union{AbstractString,Char})

A ConstantParser matching x.

parser(x)

A ConstantParser{1} matching x.

parser(x::StepRange{Char,<:Integer})

CharIn matching x.

parser(x::Pair{Symbol, P}) where P

A parser labelled with name x.first. Labels are useful in printing and logging.

See also: @with_names, with_name, log_names

parser(constant::Pair{<:ParserTypes})
map_constant(constant, parser::AbstractToken)

A parser mapping matches of parser x.first to constant x.second.

Base.convert — Function

convert(::Type{AbstractToken},x::Union{AbstractString,Char})

A ConstantParser matching x.

CombinedParsers.Regexp.@re_str — Macro

@re_str(x,flags)

Construct a ParserWithCaptures from PCRE regex syntax, such as re"^[a-z]*$", without interpolation and unescaping (except for quotation mark " which still has to be escaped). Plug-in replacement for PCRE string macro @r_str.

The regex also accepts one or more flags, listed after the ending quote, to change its behaviour:

i enables case-insensitive matching

m treats the ^ and $ tokens as matching the start and end of individual lines, as opposed to the whole string.

s allows the . modifier to match newlines.

x enables "comment mode": whitespace is ignored except when escaped with \, and # is treated as starting a comment.

a disables UCP mode (enables ASCII mode). By default \B, \b, \D, \d, \S, \s, \W, \w, etc. match based on Unicode character properties. With this option, these sequences only match ASCII characters.

xx enables "extended comment mode": whitespace in bracket character matchers are ignored.

julia> re"a|c"i
|🗄... Either
├─ [Aa] CharIn
└─ [cC] CharIn
::Char

julia> re"a+c"
🗄 Sequence
├─ a+  |> Repeat
└─ c
::Tuple{Array{Char,1},Char}

See also Regcomb, parse_options.

useful importable sub-parsers

julia> using CombinedParsers.Regexp

julia> import CombinedParsers.Regexp.whitespace_horizontal

julia> import CombinedParsers.Regexp.whitespace_maybe

julia> import CombinedParsers.Regexp.word

CombinedParsers.integer_base — Function

integer_base(base,mind=0,maxd=Repeat_max)

Parser matching a integer format on base base.

Character Matchers

CombinedParsers.Bytes — Type

Bytes{N,T} <: NIndexParser{N,T}

Fast parsing of a fixed number N of indices, reinterpret(T,match)[1] the parsed vector as T, if isbitstype, or T(match) constructor otherwise.

Provide Base.get(parser::Bytes{N,T}, sequence, till, after, i, state) where {N,T} for custom conversion.

Note

Endianness can be achieved by just mapping bswap

julia> map(bswap, Bytes(2,UInt16))([0x16,0x11])
0x1611

julia> Bytes(2,UInt16)([0x16,0x11])
0x1116

CombinedParsers.AnyChar — Type

AnyChar()

Parser matching exactly one Char, returning the character.

julia> AnyChar()
re"."

CombinedParsers.CharIn — Type

CharIn(x)

Parser matching exactly one character in x.

julia> a_z = CharIn('a':'z')
re"[a-z]"

julia> parse(a_z, "a")
'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)

julia> ac = CharIn("ac")
re"[ac]"

julia> parse(ac, "c")
'c': ASCII/Unicode U+0063 (category Ll: Letter, lowercase)

CombinedParsers.CharNotIn — Type

CharNotIn(x)

Parser matching exactly one character in x.

julia> a_z = CharNotIn('a':'z')
re"[^a-z]"

julia> ac = CharNotIn("ac")
re"[^ac]"

Base.Broadcast.broadcasted — Function

Base.broadcasted(::typeof((&)), x::CharNotIn, y::CharNotIn)

Character matchers m like Union{CharIn,CharNotIn,T}, or any type T providing a ismatch(m::T,c::Char)::Bool method represent a lazy bitarray for all characters.

other

CombinedParsers.WithMemory — Type

WithMemory(x) <: AbstractString

String wrapper with memoization of next match states for parsers at indices. Memoization is sometimes recommended as a way of improving the performance of parser combinators (like state machine optimization and compilation for regular languages).

Note

A snappy performance gain could not be demonstrated so far, probably because the costs of state memory allocation for caching are often greater than recomputing a match. If you have a case where your performance benefits with this, let me know!

```

CombinedParsers.MappedChars — Type

MappedChars(f::Function,x) <: AbstractString

String implementation lazily transforming characters. Used for fast caseless matching.

DocTestFilters = r"[0-9.]+ .s.*"

julia> p = caseless("AlsO")
🗄  |> MappingParser
├─ also
└─ lowercase
::String

julia> p("also")
"also"

julia> using BenchmarkTools;

julia> @btime match(p,"also");
  51.983 ns (2 allocations: 176 bytes)

julia> p = parser("also"); @btime match(p,"also");
  44.759 ns (2 allocations: 176 bytes)

Sequences

CombinedParsers.Sequence — Type

Sequence(;kw...)

Sequence keyword argument constructors transform the parsing into a named tuple.

Base.:* — Method

(*)(x::Any, y::AbstractToken)
(*)(x::AbstractToken, y::Any)
(*)(x::AbstractToken, y::AbstractToken)

Chain parsers in sSequence. See also @seq.

CombinedParsers.sSequence — Function

sSequence(x...)

Simplifying Sequence, flatten Sequences, remove Always lookarounds.

julia> Sequence('a',CharIn("AB")*'b')
🗄 Sequence
├─ a
└─ 🗄 Sequence
   ├─ [AB] CharIn
   └─ b
::Tuple{Char,Tuple{Char,Char}}

julia> sSequence('a',CharIn("AB")*'b')
🗄 Sequence
├─ a
├─ [AB] CharIn
└─ b
::Tuple{Char,Char,Char}

See also Sequence

CombinedParsers.@seq — Macro

@seq(x...)

Create a sequence interleaved with whitespace (horizontal or vertical). The result_type is omitting whitespace.

Either

CombinedParsers.Either — Type

Parser that tries matching the provided parsers in order, accepting the first match, and fails if all parsers fail.

This parser has no custom == and hash methods because it can recurse.

julia> match(r"a|bc","bc")
RegexMatch("bc")

julia> parse(Either("a","bc"),"bc")
"bc"

julia> parse("a" | "bc","bc")
"bc"

Base.:| — Method

(|)(x::AbstractToken, y)
(|)(x, y::AbstractToken)
(|)(x::AbstractToken, y::AbstractToken)

Operator syntax for sEither(x, y).

julia> 'a' | CharIn("AB") | "bc"
|🗄... Either
├─ a
├─ [AB] CharIn
└─ bc
::Union{Char, SubString}

Base.:| — Method

(|)(x::AbstractToken, y)
(|)(x, y::AbstractToken)
(|)(x::AbstractToken, y::AbstractToken)

Operator syntax for sEither(x, y).

julia> 'a' | CharIn("AB") | "bc"
|🗄... Either
├─ a
├─ [AB] CharIn
└─ bc
::Union{Char, SubString}

Base.:| — Method

`(|)(x::Either, T::Type)`

Return new Either with T added to result_type(x). todo: Note that the options array is kept. As a consequence push!on result will also push to x.

Base.push! — Function

Base.push!(x::Either, option)

Push option to x.options as parser tried next if x fails. Recursive parsers can be built with push! to Either.

See also pushfirst!.

Base.pushfirst! — Function

Base.pushfirst!(x::Either, option)

Push option to x.options as parser tried first, and trying x if option fails. Recursive parsers can be built with pushfirst! to Either.

See also push!.

CombinedParsers.sEither — Function

sEither(x...)

Simplifying Either, flattens nested Eithers, remove Never parsers.

julia> Either('a',CharIn("AB")|"bc")
|🗄... Either
├─ a
└─ |🗄... Either
   ├─ [AB] CharIn
   └─ bc
::Union{Char, SubString}

julia> sEither('a',CharIn("AB")|"bc")
|🗄... Either
├─ a
├─ [AB] CharIn
└─ bc
::Union{Char, SubString}

See also Either

CombinedParsers.either_result_type — Function

return tuple(statetype,resulttype)

Combining Parser

CombinedParsers.Atomic — Type

Atomic(x)

A parser matching p, and failing when required to backtrack (behaving like an atomic group in regular expressions).

CombinedParsers.FlatMap — Type

FlatMap{P,S,Q<:Function,T} <: CombinedParser{S,T}

Like Scala's fastparse FlatMap. See after

CombinedParsers.after — Function

after(right::Function,left::AbstractToken)
after(right::Function,left::AbstractToken,T::Type)

Like Scala's fastparse FlatMap

julia> saying(v) = v == "same" ? v : "different";

julia> p = after(saying, String, "same"|"but")
🗄 FlatMap
├─ |🗄... Either
│  ├─ same 
│  └─ but 
└─ saying
::String

julia> p("samesame")
"same"

julia> p("butdifferent")
"different"

Repeating

CombinedParsers.Repeat — Type

Repeat(x)
Repeat(x; min=0,max=Repeat_max)
Repeat(min::Integer, x)
Repeat(min::Integer,max::Integer, x)

Parser repeating pattern x min:max times.

julia> Repeat(2,2,'a')
a{2}  |> Repeat
::Array{Char,1}

julia> Repeat(3,'a')
a{3,}  |> Repeat
::Array{Char,1}

CombinedParsers.Lazy — Type

Lazy(x::Repeat)
Lazy(x::Optional)

Lazy x repetition matching from greedy to lazy.

julia> re"a+?"
a+?  |> Repeat |> Lazy
::Array{Char,1}

julia> re"a??"
a?? |missing |> Lazy
::Union{Missing, Char}

CombinedParsers.Optional — Type

Optional(parser;default=defaultvalue(result_type(parser)))

Parser that always succeeds. If parser succeeds, return result of parser with curser behind match. If parser does not succeed, return default with curser unchanged.

julia> match(r"a?","b")
RegexMatch("")

julia> parse(Optional("a", default=42),"b")
42

Base.:| — Method

(|)(x::AbstractToken{T}, default::Union{T,Missing})

Operator syntax for Optional(x, default=default).

julia> parser("abc") | "nothing"
|🗄... Either
├─ abc
└─ nothing
::SubString

julia> parser("abc") | missing
abc? |missing
::Union{Missing, SubString}

CombinedParsers.Repeat1 — Function

Repeat1(x)

Parser repeating pattern x one time or more.

Repeat1(f::Function,a...)

Abbreviation for map(f,Repeat1(a...)).

CombinedParsers.Repeat_stop — Function

Repeat_stop(p,stop)
Repeat_stop(p,stop; min=0, max=Repeat_max)

Repeat p until stop (NegativeLookahead), not matching stop. Sets cursor before stop. Tries min:max times Returns results of p.

julia> p = Repeat_stop(AnyChar(),'b') * AnyChar()
🗄 Sequence
├─ 🗄* Sequence[2] |> Repeat
│  ├─ (?!🗄) NegativeLookahead
│  │  └─ b
│  └─ . AnyChar
└─ . AnyChar
::Tuple{Array{Char,1},Char}

julia> parse(p,"acbX")
(['a', 'c'], 'b')

See also NegativeLookahead

CombinedParsers.Repeat_until — Function

Repeat_until(p,until, with_until=false; wrap=identity, min=0, max=Repeat_max)

Repeat p until stop (with Repeat_stop). and set point after stop.

Return a Vector{result_type(p)} if wrap_until==false, otherwise a Tuple{Vector{result_type(p)},result_type(until)}.

To transform the Repeat_stop(p) parser head, provide a function(::Vector{result_type(p)}) in wrap keyword argument, e.g.

julia> p = Repeat_until(AnyChar(),'b') * AnyChar()
🗄 Sequence
├─ 🗄 Sequence[1]
│  ├─ (?>🗄*) Sequence[2] |> Repeat |> Atomic
│  │  ├─ (?!🗄) NegativeLookahead
│  │  │  └─ b
│  │  └─ . AnyChar
│  └─ b
└─ . AnyChar
::Tuple{Array{Char,1},Char}

julia> parse(p,"acbX")
(['a', 'c'], 'X')

julia> parse(Repeat_until(AnyChar(),'b';wrap=JoinSubstring),"acbX")
"ac"

See also NegativeLookahead

Base.join — Function

Base.join(x::Repeat,delim, infix=:skip)

Parser matching repeated x.parser separated by delim.

julia> parse(join(Repeat(AnyChar()),','),"a,b,c")
3-element Array{Char,1}:
 'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)
 'b': ASCII/Unicode U+0062 (category Ll: Letter, lowercase)
 'c': ASCII/Unicode U+0063 (category Ll: Letter, lowercase)

julia> parse(join(Repeat(AnyChar()),',';infix=:prefix),"a,b,c")
('a', [(',', 'b'), (',', 'c')])

julia> parse(join(Repeat(AnyChar()),',';infix=:suffix),"a,b,c")
([('a', ','), ('b', ',')], 'c')

Base.join(x::AbstractToken,delim)

Shorthand for join(Repeat(x),delim).

Base.join(f::Function, x::AbstractToken, delim)

Shorthand for map(f,join(x,delim)).

Logging and Side-Effects

CombinedParsers.@syntax — Macro

Convenience macro @syntax name = expr for defining a CombinedParser name=expr and custom parsing macro @name_str.

With @syntax for name in either; expr; end the defined parser is pushfirst! to either. If either is undefined, it will be created. If either == :text || either == Symbol(:) the parser will be added to CombinedParser_globals variable in your module.

DocTestFilters = r"map\(#[^)]+\)"

julia> @syntax for german_street_address in street_address
            Sequence(:street => !Repeat(AnyChar()),
                 " ",
                 :no =>Numeric(Int))
       end
🗄 Sequence |> map(ntuple) |> with_name(:german_street_address)
├─ .* AnyChar |> Repeat |> ! |> with_name(:street)
├─ \  
└─ Int64  |> with_name(:no)
::NamedTuple{(:street, :no),Tuple{SubString,Int64}}

julia> german_street_address"Some Avenue 42"
NamedTuple{(:street, :no),Tuple{SubString,Int64}}(("Some Avenue", 42))

julia> @syntax for us_street_address in street_address
            Sequence(:no =>Numeric(Int),
                     " ",
                     :street => !Repeat(AnyChar()))
       end
🗄 Sequence |> map(ntuple) |> with_name(:us_street_address)
├─ Int64  |> with_name(:no)
├─ \  
└─ .* AnyChar |> Repeat |> ! |> with_name(:street)
::NamedTuple{(:no, :street),Tuple{Int64,SubString}}

julia> street_address"50 Oakland Ave"
NamedTuple{(:no, :street),Tuple{Int64,SubString}}((50, "Oakland Ave"))

CombinedParsers.NamedParser — Type

NamedParser{P,S,T} <: WrappedParser{P,S,T}

Struct with

    name::Symbol
    parser::P
    doc::String

CombinedParsers.with_name — Function

with_name(name::Symbol,x; doc="")

A parser labelled with name. Labels are useful in printing and logging.

See also: @with_names, with_name, log_names

CombinedParsers.@with_names — Macro

@with_names

Sets names of parsers within begin/end block to match the variables they are asigned to.

so, for example

julia> @with_names foo = AnyChar()
. AnyChar |> with_name(:foo)
::Char

julia> parse(log_names(foo),"ab")
   match foo@1-2: ab
                  ^
'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)

See also log_names(parser), @syntax.

CombinedParsers.log_names — Function

log_names(x,names=true; exclude=nothing)

Rebuild parser replacing NamedParser instances with with_log parsers. Log all NamedParser instanses if names==true or name in names and not name in exclude.

See also: with_log, deepmap_parser

CombinedParsers.with_log — Function

with_log(s::AbstractString,p, delta=5;nomatch=false)

Log matching process of parser p, displaying delta characters left of and right of match.

If nomatch==true, also log when parser does not match.

See also: log_names, with_effect

CombinedParsers.with_effect — Function

with_effect(f::Function,p,a...)

Call f(sequence,before_i,after_i,state,a...) if p matches, f(sequence,before_i,before_i,nothing,a...) otherwise.

Assertions

CombinedParsers.AtStart — Type

AtStart()

Parser succeding if and only if at index 1 with result_type AtStart.

julia> AtStart()
re"^"

CombinedParsers.AtEnd — Type

AtEnd()

Parser succeding if and only if at last index with result_type AtEnd.

julia> AtEnd()
re"$"

CombinedParsers.Always — Type

Always()

Assertion parser matching always and not consuming any input. Returns Always().

julia> Always()
re""

CombinedParsers.Never — Type

Never()

Assertion parser matching never.

julia> Never()
re"(*FAIL)"

Look behind

Base.:/ — Method

(/)(x::ParserTypes, y::ParserTypes)

Sequence(PositiveLookbehind(x),y)

julia> match("is "/"match", "no match is match").start
13

Note

This syntax is reviewed and I your appreciate your comments!

CombinedParsers.Lookbehind — Function

Lookbehind(does_match::Bool, p)

PositiveLookbehind if does_match==true, NegativeLookbehind otherwise.

CombinedParsers.PositiveLookbehind — Type

PositiveLookbehind(parser)

Parser that succeeds if and only if parser succeeds before cursor. Consumes no input. The match is returned. Useful for checks like "must be preceded by parser, don't consume its match".

CombinedParsers.NegativeLookbehind — Type

NegativeLookbehind(parser)

Parser that succeeds if and only if parser does not succeed before cursor. Consumes no input. nothing is returned as match. Useful for checks like "must not be preceded by parser, don't consume its match".

julia> la=NegativeLookbehind("keep")
re"(?<!keep)"

julia> parse("peek"*la,"peek")
("peek", re"(?<!keep)")

CombinedParsers.reverse_index — Function

reverse_index(x::Reverse,i)

Return corresponding index in unreversed String x.lastindex-i+1. Cap at 0:lastindex+1. (can be optimized maybe)

Look ahead

CombinedParsers.Lookahead — Function

Lookahead(does_match::Bool, p)

PositiveLookahead if does_match==true, NegativeLookahead otherwise.

CombinedParsers.PositiveLookahead — Type

PositiveLookahead(parser)

Parser that succeeds if and only if parser succeeds, but consumes no input. The match is returned. Useful for checks like "must be followed by parser, but don't consume its match".

julia> la=PositiveLookahead("peek")
re"(?=peek)"

julia> parse(la*AnyChar(),"peek")
("peek", 'p')

CombinedParsers.NegativeLookahead — Type

NegativeLookahead(parser)

Parser that succeeds if and only if parser does not succeed, but consumes no input. parser is returned as match. Useful for checks like "must not be followed by parser, don't consume its match".

julia> la = NegativeLookahead("peek")
re"(?!peek)"

julia> parse(la*AnyChar(),"seek")
(re"(?!peek)", 's')