CombinedParsers.BNF — Module

(Extended) Backus-Naur Form CombinedParser

Defining a EBNF parser can be done with the CombinedParsers.BNF.ebnf string macro. substitute is used to combine parts of the definition.

Warn

Left recursion is not yet supported (will lead to a stack overflow).

CombinedParsers.BNF.ebnf — Constant

ebnf

Parser to create a CombinedParser from EBNF syntax:

julia> p = ebnf"""
       digit excluding zero = "1" | "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9" ;
       digit                = "0" | digit excluding zero ;
       natural number       = digit excluding zero, { digit } ;
       integer              = "0" | [ "-" ], natural number ;
       """
|🗄 Either
├─ |🗄 Either |> with_name(:integer)
│  ├─ 0 
│  └─ 🗄 Sequence
│     ├─ \-? |
│     └─ 🗄 Sequence |> with_name(:natural number) # branches hidden
├─ 🗄 Sequence |> with_name(:natural number)
│  ├─ |🗄 Either |> with_name(:digit excluding zero) # branches hidden
│  └─ |🗄* Either |> with_name(:digit) |> Repeat
│     ├─ 0 
│     └─ |🗄 Either |> with_name(:digit excluding zero) # branches hidden
├─ |🗄 Either |> with_name(:digit)
│  ├─ 0 
│  └─ |🗄 Either |> with_name(:digit excluding zero) # branches hidden
└─ |🗄 Either |> with_name(:digit excluding zero)
   ├─ 1 
   ├─ 2 
   ├─ 3 
   ├─ 4 
   ├─ 5 
   ├─ 6 
   ├─ 7 
   ├─ 8 
   └─ 9 
::Union{SubString{String}, Tuple{SubString{String}, Vector{SubString{String}}}, Tuple{AbstractString, Tuple{SubString{String}, Vector{SubString{String}}}}}

julia> p[:integer]("42")
("", ("4", SubString{String}["2"]))

A (too complicated) result type is derived implicitly. You can map transform results of parts of a EBNF parser with the deepmap function:

julia> deepmap(MatchedSubSequence, p, :integer)[:integer]("42")
"42"

Note

I want to support more BNF variants. Contributions of test cases are welcome! A EBNF Syntax draft built from Wikimedia Ebnf-syntax-diagram.

Warn

Left recursion is not yet supported (will lead to a stack overflow).

CombinedParsers.BNF.defining_symbol — Constant

Supports BNF and EBNF variants

julia> CombinedParsers.BNF.defining_symbol

CombinedParsers.BNF.concatenate_symbol — Constant

Supports BNF and EBNF variants

julia> CombinedParsers.BNF.concatenate_symbol