CombinedParsers.jl Documentation

A package for combining parsers and transforming strings into julia types.

Compose parsers with the functional parser combinator paradigm, utilize Julia's type inference for transformations, log conveniently for debugging, and let Julia compile your parser for performance.

Note

CombinedParsers.jl is currently an α release. The first official released is prepared for JuliaCon2020.

Package Features

Speed

write parsers faster than Base.PCRE, optimized by the Julia compiler for parametric parser and state types.

@generated functions, trie-based scanning (example), compile with your custom parsing algorithm (example)

(planned: memoization, lazy transformations)

Simplicity

Clear @syntax integrates map transformations with Julia result_type inference.

Define without redundancy: parser, memory representation, and instance construction. When solely the parser is defined, Julia infers result_type(parser) and defines memory layout, and constructors are compiled for the parsing state from Transformations.

AbstractTrees.jl interface provides clearly layed out printing in the REPL. with_log provides colored logging of the parsing with_names.

Interoperability

TextParse.jl: existing TextParse.AbstractToken implementations can be used with CombinedParsers. CombinedParser provide TextParse.tryparsenext and can be used e.g. in CSV.jl.

Pure Julia regular expression parsers are provided with the @re_str macro, a plug-in replacement for Base.@r_str. Tested on the PCRE pattern test set.

Generality

All valid parsings can be Base.iterated lazily.

Higher-order parsers depending on the parsing state allow for not context-free parsers (after).

can process UTF8 strings or any sequence type supporting getindex, nextind, prevind methods.

Getting started

The Overview provides a tutorial explaining how to get started using CombinedParsers.

The User guide provides a summary of CombinedParsers types and constructors. Some examples of packages using CombinedParsers can be found on the Examples page. See the Index for the complete list of documented functions and types.

Install with

] add https://github.com/gkappler/CombinedParsers.jl

Example: rational numbers arithmetics

Parsing is reading and transforming a sequence of characters. CombinedParsers provides constructors to combine parsers and transform (sub-)parsings arbitrarily with julia syntax. Combinator constructors are discussed in the user guide.

using CombinedParsers
using TextParse

This example reads and evaluates arithmetical terms for rational numbers. The following defines an evaluating parser for rational number terms as sequences of subterms interleaved with operators.

Sub-terms are Either fast TextParse.Numeric(Int) integer numbers, converted to Rational{Int}, or a subterm is written as parentheses around a nested term:

@syntax subterm = Either{Rational{Int}}(TextParse.Numeric(Int));
@syntax for parenthesis in subterm
    mult         = evaluate |> join(subterm, CharIn("*/"), infix=:prefix )
    @syntax term = evaluate |> join(mult,    CharIn("+-"), infix=:prefix )
    Sequence(2,'(',term,')')
end;

For parsing, @syntax registers a @term_string macro for parsing and transforming.

julia> term"(1+2)/5"
3//5

julia> # The defined `CombinedParser` `term` function
       # provides optional logging of the parsing process.
       term("1/((1+2)*4+3*(5*2))",log = [:parenthesis])
   match parenthesis: 1/((1+2)*4+3*(
                         ^___^
   match parenthesis: *4+3*(5*2))
                           ^___^
   match parenthesis: 1/((1+2)*4+3*(5*2))
                        ^_______________^
1//42

Is every rational answer ultimately the inverse of a universal question in life?

This @syntax definition in 5,5 lines is sufficient for parsing and evaluating arithmetics: Base.join(x,infix; infix=:prefix) is shorthand for x*Repeat( infix * x ), and f |> parser is shorthand for map(f,parser).

Note: The evaluate function definition is detailed in the full example.

julia> evaluate( (0, [ ('+',1), ('-',2) ]) )
-1//1

julia> evaluate( (1, [ ('*',4), ('/',3) ]) )
4//3

Design

The arithmetic parser can be defined without redundancy. The CombinedParsers design

infers the Julia data types of the parsing domain with result_type,

constructors are compiled for the parsing states,

is fast due to Julia parametric types, and compiler optimizations with generated functions,

is composable and optimizable with Julia method dispatch,

provides flexible public API for parsing, matching, and iteration.

This design is used in several packages of format-parsers transforming texts into nested julia representations.

OrgmodeParser.jl is a CombinedParser for parsing main org mode syntax, representing org files within Julia.

WikitextParser.jl is a CombinedParser for parsing wikitext syntax, quite comprehensibly and representing Wikipedia articles within Julia.

CombinedParserTools.jl is currently more or less my own workspace to provide a set of re-useable parsers.

If you want to work with any of these open source packages, I will gladly provide professional support. If you are writing your own recursive CombinedParser and seek inspiration, you might find these comprehensive examples interesting. (pre-\alpha, so beware, dragons!)

Making Julia parametric types central for the parser design allows equal automation of the data pipeline after parsing!

fast db-indexing of text streams (e.g. logging): If you need support indexing logging streams into a (SQL-)Database, the (currently) proprietary TypeGraphs.jl provides CombinedParsers plug and play: Table schemas are infered from your parser.

fast HTTP-serving of parsed data: If you need support with a parsing server-client infrastructure, the (currently) proprietary GraphQLAlchemy.jl provides CombinedParsers plug and play: GraphQL schemas and resolver are infered from your parser.

fast out-of core data science/AI on your parsed data: If you need support with storing parsed data in optimized memory-mapped JuliaDB, TypeDB.jl provides CombinedParsers plug and play.

fast scientific measurements in a data graph: FilingForest IA.jl provides CombinedParsers plug and play: even for recursively nested data.

All (currently) proprietary packages are default-over-configuration for fast integration, and are in active development.

Optimization Strategy

CombinedParsers.jl is tested against the C PCRE2 library testset. This strategy also allows for efficient benchmarking of code optimizations on many patterns, and runtime comparison with C PCRE2. C PCRE2 optimized is among the fastest regex libraries (second behind Rust, running mariomka's benchmark will position CombinedParser among its competition. Explorations for optimization are in git branches.

All benchmarks are wrong, but some are useful - Szilard, benchm-ml

Compliance with the PCRE test set

Test Overview

Performance Overview:

Benchmark ratios histogram:

Unsupported

The package is still young, and optimization is ongoing. If you are interested in and able to dive deeper into the Julia memory layout and compiler, I would gladly collaborate on further optimizations:

String layout: Parsing requires repeated Char comparisons. In UTF8, frequent characters are encoded shorter (8 bit), rare have longer codes. For this reason, in Julia String indices are not consecutive and transversal requires using infamous nextind and prevind. Profiling:

nextind and prevind comsume considerable time. Could be cached/memoized?

CombinedParsers currently operates on the result of getindex(::String,index)::Char (technically on iterate(::String,index)::Tuple{Char,Int}). Could matching use the raw byte representation directly?

Macros: make all iteration @generated functions using expressions generated by a dispatched iterate_expression that can be used in a macro @iterate to generate an unrolled/unnested iteration code. (Profiling hints that function calls do hardly contribute to runtime.)

Acknowledgements

This package is enabled only due to the Julia's compiler and superior type system. Thankfully: a really concise language for powerful computing!

I am thankful for contributions and inspiration from many great packages:

TextParse.jl

A bunch of fast text parsing tools, used in CSV.jl

CombinedParsers composes with fast TextParse.jl both ways because CombinedParser <: TextParse.AbstractToken and by providing a method for TextParse.tryparsenext, (leveraging the supreme Julia compiler, type and package architecture).

If you seek support with a dates parser example, please contact me.

If you seek support with a CSV example, please contact me (e.g. address text field parsing).

Dates.tryparsenext — Function

TextParse.tryparsenext(x::CombinedParser,str,i,till,opts=TextParse.default_opts)

TextParse.jl integrates with CombinedParsers.jl both ways.

DocTestFilters = r"map\(.+\)"

julia> p = ("Number:" * Repeat(' ') * TextParse.Numeric(Int))[3]
🗄 Sequence |> map(IndexAt(3))
├─ Number\:
├─ \ *  |> Repeat
└─ <Int64>
::Int64

julia> parse(p, "Number:    42")
42

julia> TextParse.tryparsenext(p, "Number:    42")
(Nullable{Int64}(42), 14)

CombinedParsers builds on

Tries.jl is the abstract implementation of the fast prefix-tree matching in CombinedParsers (see docs)

AbstractTrees.jl interface provides clearly layed out printing in the REPL. with_log provides colored logging of the parsing with_names.

Inspirations

The work was strongly inspired by the great Scala fastparse package, and also the elm parser.

Parsers.jl, a collection of parsers for date and primitive types, inspired the parse methods.

Automa.jl, a Julia package for text validation, parsing, and tokenizing based on state machine compiler. The package compiles deterministic finite automata. (Currently there is no inter-operation possible, because in Automa processing of parsed tokens is done with actions and UTF8 support is lacking).

ParserCombinator.jl was a great inspiration. Yet I decided for a new design with a focus on transformations and type inference with parametric types, instead of basing this work off ParserCombinator, written before 2016 (and fixed for Julia 1.0 in 2018). CombinedParsers integrates into the Julia 1.0 Iteration API, small Union{Nothing,T} where T types instead of using Nullables, compiler optimizations and generated functions. I want to provide benchmarks comparisons with ParserCombinator.jl.

Outline

Manual

Overview

ParseMatch

Parsing

Iterating

Performance

Transformations

User Guide

Basics

Character Sets

Sequence

Either

Repeat

Optional

Lazy repetitions and optional parsers

Assertions

AtStart() and AtEnd()

Looking around

Atomic groups

A fast Trie-based parser for a collection of literal Strings.

Next steps

Examples

What is Parsing?

Names

Addresses

Person's address data

Number lists (wikitext references)

Number ranges

Numbers

Joining numbers and ranges

Inclusion in a wikitext parser

PCRE papercuts when parsing number sequences

Regular Expressions CombinedParser

What is a regular expression?

Characters and Escaped Characters

Repeated patterns

Number of Repetitions: Transforming a Parsing

Either

Repeatable patterns, Optional, Lazy and Atomic

Sequences and Alternations

Captures

Lookahead and Lookbehind

Regular Expression Brackets

Palindromes<:CombinedParser: a Tutorial for writing your combinable Parser

1. Regular Expression

Tree display of regex

Regular Expression performance

2. A non-word skipping Palindrome<:CombinedParser

Parsing strategy

Prerequisite: Skipping whitespace

Subtyping <: CombinedParser{STATE,RESULT}.

Matching: CombinedParsers._iterate

Base.prevind and Base.nextind

match and get

Iterating through matches

Iteration of smaller Sub-palindromes

Performance Optimization

Padding and combining

Next...

JSON

Library API

Public Documentation

Matching

Parsing

Typed Transformation Parsers

Parser Constructors

useful importable sub-parsers

Character Matchers

Combining Parser

Repeating

Logging and Side-Effects

Assertions

Internal API

Iterating Parsings

Internal Types

Abstract Parsers

Wrapped Parsers

States

Printing

Rewriting Parsers

Index

CombinedParsers.Lookahead

CombinedParsers.Lookbehind

CombinedParsers.Repeat1

CombinedParsers.Repeat_stop

CombinedParsers.Repeat_until

CombinedParsers.either_types

CombinedParsers.log_names

CombinedParsers.map_at

CombinedParsers.parser

CombinedParsers.result_type

CombinedParsers.sEither

CombinedParsers.sSequence

CombinedParsers.state_type

CombinedParsers.tryparse_pos

CombinedParsers.with_effect

CombinedParsers.with_log

CombinedParsers.with_name

CombinedParsers.Always

CombinedParsers.AnyChar

CombinedParsers.AtEnd

CombinedParsers.AtStart

CombinedParsers.Atomic

CombinedParsers.CharIn

CombinedParsers.CharNotIn

CombinedParsers.Constant

CombinedParsers.Either

CombinedParsers.JoinSubstring

CombinedParsers.Lazy

CombinedParsers.MatchesIterator

CombinedParsers.NegativeLookahead

CombinedParsers.NegativeLookbehind

CombinedParsers.Never

CombinedParsers.Optional

CombinedParsers.ParseMatch

CombinedParsers.PositiveLookahead

CombinedParsers.PositiveLookbehind

CombinedParsers.Repeat

CombinedParsers.Sequence

CombinedParsers.Regexp.integer_base

CombinedParsers._ismatch

CombinedParsers._iterate

CombinedParsers.deepmap_parser

CombinedParsers.ismatch

CombinedParsers.print_constructor

CombinedParsers.regex_inner

CombinedParsers.regex_prefix

CombinedParsers.regex_string

CombinedParsers.regex_suffix

CombinedParsers.tuple_pos

CombinedParsers.tuple_state

CombinedParsers.CombinedParser

CombinedParsers.ConstantParser

CombinedParsers.FilterParser

CombinedParsers.LeafParser

CombinedParsers.LookAround

CombinedParsers.MatchState

CombinedParsers.MemoTreeChildren

CombinedParsers.NCodeunitsState

CombinedParsers.NIndexParser

CombinedParsers.None

CombinedParsers.Transformation

CombinedParsers.WrappedParser

CombinedParsers.Regexp.NoDict