CombinedParsers.jl Documentation

A package for combining parsers and transforming strings into julia types.

Compose parsers with the functional parser combinator paradigm, utilize Julia's type inferrence for transformations, log conveniently for debugging, and let Julia compile your parser for good performance.

Note

CombinedParsers.jl is currently an α release. The first official released is prepared for JuliaCon2020.

Package Features

Clear syntax integrates grammar and transformations with Julia type inference.

Higher-order parsers depending on the parsing state allow for not context-free parsers.

All valid parsings can be iterated lazily.

Interoperable with TextParse.jl: existing TextParse.AbstractToken implementations can be used with CombinedParsers. CombinedParser provide TextParse.tryparsenext and can be used e.g. in CSV.jl.

Parametric parser and state types enable Julia compiler optimizations.

Compiled regular expression parsers in pure julia are provided with the re_str macro.

AbstractTrees.jl interface provides colored and clearly layed out printing in the REPL.

Convenient logging of the parsing process with NamedParsers and SideeffectParsers.

CombinedParsers generalize from strings to parsing any type supporting getindex, nextind, prevind methods.

Getting started

Install with

] add https://github.com/gkappler/CombinedParsers.jl

The Overview provides a tutorial explaining how to get started using CombinedParsers. The User guide provides a summary of CombinedParsers types. Some examples of packages using CombinedParsers can be found on the Examples page. See the Index for the complete list of documented functions and types.

Writing Parsers

CombinedParsers provides constructors to combine parsers and transform (sub-)parsings arbitrarily with julia syntax. Combinator constructors are discussed in the user manual.

A term expression has

using CombinedParsers
@syntax subterm = Either{Rational{Int}}()

# A integer subterm can be parsed with fast TextParse.jl parsers:
using TextParse
@syntax for number in subterm
    map(Rational{Int},TextParse.Numeric(Int))
end

# Subterms can use algebraic operators `+-*/` that will be evaluated
function evaluate( (start, operation_values) )
    aggregated_value::Rational{Int} = start
    for (op,val) in operation_values
        aggregated_value = eval( Expr(:call, Symbol(op),
			              aggregated_value, val
			              ))
    end
    return aggregated_value
end

# Terms are sequences of subterms interleaved with operators.
# CombinedParsers.Sequence can be built with operator `*` like Julia string concatenation.
rational_products = Sequence(
	evaluate,
	subterm, Repeat(CharIn("*/"), subterm )
)

@syntax term = map(
    evaluate,
    rational_products*Repeat(CharIn("+-")*rational_products)
);

# A subterm can also be a nested term in parenthesis
@syntax for parenthesis in subterm
    Sequence(2,"(",term,")")
end

@syntax registers @term_string macro for parsing.

julia> term"(1+2)/5"
3//5

The defined CombinedParser term can be used as a function for colorful logging of the parsing process.

julia> term("1/((1+2)*4+3*(5*2))",log = [:parenthesis])
   match parenthesis: 1/((1+2)*4+3*(
                         ^___^
   match parenthesis: *4+3*(5*2))
                           ^___^
   match parenthesis: 1/((1+2)*4+3*(5*2))
                        ^_______________^
1//42

The parser representation can be printed as a tree

julia> term
🗄 Sequence |> map(#43) |> with_name(:term)
├─ 🗄 Sequence |> map(#43)
│  ├─ |🗄... Either |> with_name(:subterm)
│  │  ├─ 🗄 Sequence |> map(#51) |> with_name(:parenthesis)
│  │  │  ├─ \(
│  │  │  ├─ 🗄 Sequence |> map(#43) |> with_name(:term)
│  │  │  │  ├─ 🗄 Sequence |> map(#43)
│  │  │  │  └─ 🗄* Sequence |> Repeat
│  │  │  └─ \)
│  │  └─ TextParse.Numeric{Int64} TextParse.Numeric |> map(#45) |> with_name(:number)
│  └─ 🗄* Sequence |> Repeat
│     ├─ [\*/] CharIn
│     └─ |🗄... Either |> with_name(:subterm)
│        ├─ 🗄 Sequence |> map(#51) |> with_name(:parenthesis)
│        │  ├─ \(
│        │  ├─ 🗄 Sequence |> map(#43) |> with_name(:term)
│        │  └─ \)
│        └─ TextParse.Numeric{Int64} TextParse.Numeric |> map(#45) |> with_name(:number)
└─ 🗄* Sequence |> Repeat
   ├─ [\+\-] CharIn
   └─ 🗄 Sequence |> map(#43)
      ├─ |🗄... Either |> with_name(:subterm)
      │  ├─ 🗄 Sequence |> map(#51) |> with_name(:parenthesis)
      │  │  ├─ \(
      │  │  ├─ 🗄 Sequence |> map(#43) |> with_name(:term)
      │  │  └─ \)
      │  └─ TextParse.Numeric{Int64} TextParse.Numeric |> map(#45) |> with_name(:number)
      └─ 🗄* Sequence |> Repeat
         ├─ [\*/] CharIn
         └─ |🗄... Either |> with_name(:subterm)
            ├─ 🗄 Sequence |> map(#51) |> with_name(:parenthesis)
            └─ TextParse.Numeric{Int64} TextParse.Numeric |> map(#45) |> with_name(:number)
::Rational{Int64}

Is every rational answer ultimately the inverse of a universal question in life?

Parsing has been a painpoint in all of my work as a developer and a researcher with a focus on text.

Examples

JSON

Parsing Regular Expressions and constructing an equivalent CombinedParser

Characters and Escaped Characters

Transforming a Parsing

Repeated patterns

Number of Repetitions

Either

Repeatable patterns, Optional, Lazy and Atomic

Sequences and Alternations

Captures

Lookahead and Lookbehind

Regular Expression Brackets

Optimization Strategy

CombinedParsers.jl is tested and benchmarked against the PCRE C library testset.

Compliance with the PCRE test set

Test Overview

Performance Overview:

Benchmark ratios histogram:

Unsupported

This strategy allows for efficient benchmarking of code optimizations on many Regex and other syntax patterns. Explorations for optimization are in git branches:

Acknowledgements

I am thankful for contributions and inspiration from many great packages:

The work was inspired by Scala fastparse package and the Julia parsing packages

Parsers.jl

A collection of parsers for date and primitive types.

TextParse.jl

A bunch of fast text parsing tools.

used in CSV.jl

Nullables.jl

I aim at making CombinedParsers compose with [TextParse.jl](https://github.com/queryverse/TextParse.jl) both ways by type CombinedParser <: TextParse.AbstractToken and providing a method for TextParse.tryparsenext. I will contribute soon a CSV example with address text field parsing.

Dates.tryparsenext — Function

TextParse.tryparsenext(x::CombinedParser,str,i,till,opts=TextParse.default_opts)

TextParse.jl integrates with CombinedParsers.jl both ways.

julia> p = ("Number:" * Repeat(' ') * TextParse.Numeric(Int))[3]
🗄 Sequence |> map(#37)
├─ Number\:
├─ \ *  |> Repeat
└─ <Int64>
::Int64

julia> parse(p, "Number:    42")
42

julia> TextParse.tryparsenext(p, "Number:    42")
(Nullable{Int64}(42), 14)

Automa.jl

A Julia package for text validation, parsing, and tokenizing based on state machine compiler. The package compiles a parser deterministic finite automaton. Processing of parsed tokens is done with actions. No UTF8 support.

ParserCombinator.jl

old source base (pre 2016, fixed for Julia 1.0 in 2018)

using Nullables

no iterator API

performance

mutable matcher types

matcher types not parametric

Library Outline

Public Documentation

CombinedParsers

CombinedParsers.Regexp

Parsing, Logging and Side-Effects

Transformations

Parser Constructors

Character Matchers

Combining Parser

Repeating

Assertions

Printing

Internal API

Iterating Parsings

Rewriting Parsers

Parsing Options

Internal Types

Abstract Parser Types

Regular Expression Types

Index

CombinedParsers.ParserTypes

CombinedParsers.CombinedParser

CombinedParsers.ConstantParser

CombinedParsers.LeafParser

CombinedParsers.LookAround

CombinedParsers.NIndexParser

CombinedParsers.Regexp.Backreference

CombinedParsers.Regexp.Capture

CombinedParsers.Regexp.Conditional

CombinedParsers.Regexp.DupSubpatternNumbers

CombinedParsers.Regexp.FilterOptions

CombinedParsers.Regexp.OnOptionsParser

CombinedParsers.Regexp.ParseMatch

CombinedParsers.Regexp.ParserOptions

CombinedParsers.Regexp.ParserWithCaptures

CombinedParsers.Regexp.SequenceWithCaptures

CombinedParsers.Regexp.Subroutine

CombinedParsers.Regexp.WithOptions

CombinedParsers.Transformation

CombinedParsers.WrappedParser

Base.get

CombinedParsers._iterate

CombinedParsers.deepmap_parser

CombinedParsers.ismatch