Internal API

Iterating Parsings

CombinedParsers.MatchesIterator — Type

Iterator type for match_all and parse_all.

CombinedParsers.parsematch_tuple — Function

parsematch_tuple(m,start,state)

ParseMatch iteration has the first match as iterator, the last match as a state. (Turned out to be fastest.)

Base.prevind — Function

Base.prevind(str,i::Int,parser,state)

Return the index before the state match ending before position i.

Note

I am in doubt whether this qualifies as type piracy because I provide an outside method for outside types. I thought it might not, because the differentiating two extra arguments. If you have an opinion, please let me know on GitHub.

Base.nextind — Function

Base.nextind(str,i::Int,parser,state)

Return the index after the state match at i.

Note

I am in doubt whether this qualifies as type piracy because I provide an outside method for outside types. I thought it might not, because the differentiating two extra arguments. If you have an opinion, please let me know on GitHub.

CombinedParsers.ismatch — Function

" ismatch(c::Char,p::Union{Function,Steprange,Set,UnicodeClass})::Bool

checks is a character matches a pattern.

CombinedParsers._ismatch — Function

_ismatch(x::Char, set::Union{Tuple,Vector})::Bool

_ismatch(x,set...) respects boolean logic:

Example:

julia> p = CharNotIn(CharNotIn("ab"));

julia> parse(p,"a")
'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)

CombinedParsers._iterate — Function

_iterate(p::ParserWithCaptures, sequence::SequenceWithCaptures,a...)

Base.empty!(sequence) before iteration.

_iterate(parser, sequence, till::Int, posi::Int[, nothing])

Dispatches to _iterate(parser, sequence,till,posi,posi,nothing) to retrieve first match, or nothing.

_iterate(parser, sequence, till, posi, next_i, states)

Note: next_i is the index in sequence after parser match according to state (and not the start of the match), such that start_index(sequence,after,parser,state) returns the start of the matching subsequence, and sequence[startindex(sequence,after,parser,state):prevind(sequence,nexti)] is the matched subsequence.

_iterate(p::Trie{Char}, str, till, posi, next_i, ::Nothing)

Match char path in p greedily, recording shorter matches in state.

CombinedParsers.tuple_pos — Function

tuple_pos(pos_state::Tuple)

pos_state[1] is position after match in tuple returned by _iterate.

CombinedParsers.tuple_state — Function

tuple_state(pos_state::Tuple)

pos_state[2] is state of match in tuple returned by _iterate.

Internal Types

Abstract Parsers

CombinedParsers.ParserTypes — Constant

Julia types that provide CombinedParser methods resulttype, statetype, _iterate, get, nextind, prevind.

CombinedParsers.CombinedParser — Type

CombinedParser{T,S} <: AbstractToken{T}

Abstract parser type for parsers returning matches transformed to ::T and state::S.

CombinedParsers.LeafParser — Type

LeafParser{T} <: AbstractToken{T}

Abstract parser type for parsers that have no sub-parser. Used for dispatch in deepmap_parser

Wrapped Parsers

CombinedParsers.FilterParser — Type

A parser succeeds ony if

the wrapped parser succeeds

and a predicate function state_filter(sequence, till, posi, r...) returns true the after,state = r tuple.

CombinedParsers.ConstantParser — Type

Wrapper for stepping with ncodeunit length.

julia> parser("constant") isa CombinedParsers.ConstantParser
true

julia> parser('c') isa CombinedParsers.ConstantParser
true

julia> parser(1) isa CombinedParsers.ConstantParser
true

CombinedParsers.NIndexParser — Type

Abstract type for stepping with previndex/nextindex, accounting for ncodeunit length of chars at point.

CombinedParsers.WrappedParser — Type

Abstract type for parser wrappers, providing default methods

CombinedParsers.Transformation — Type

Transformation{T}(f::Function, p_) where {T}
Base.map(f::Function, Tc::Type, p::ParserTypes, a...)
Base.map(f::Function, p::ParserTypes, a...)

Parser transforming result of a wrapped parser. a... is passed as additional arguments to f (at front .

CombinedParsers.LookAround — Type

Parsers that do not consume any input can inherit this type.

States

CombinedParsers.MatchState — Type

State object for a match that is defined by the parser, sequence and position.

CombinedParsers.None — Type

State type for skipped optional. (Missing was breaking julia).

CombinedParsers.NCodeunitsState — Type

State object representing ncodeunits explicitely with state of match for prevind, nextind to improve performance. nc::Int state::S

See also MatchState, prevind, nextind.

Printing

CombinedParsers.regex_string — Function

regex_string(x::CombinedParser)

regex_prefix(x)*regex_inner(x)*regex_suffix(x)

CombinedParsers.regex_prefix — Function

regex_prefix(x)

Prefix printed in parser tree node.

CombinedParsers.regex_inner — Function

regex_inner(x::AbstractToken)

Regex representation of x. See regex_string

CombinedParsers.regex_suffix — Function

regex_suffix(x)

Suffix printed in parser tree node.

CombinedParsers.print_constructor — Function

print_constructor(io::IO,x)

Print constructor pipeline in parser tree node.

CombinedParsers.MemoTreeChildren — Type

decurse recursive patterns

Rewriting Parsers

CombinedParsers.deepmap_parser — Function

deepmap_parser(f::typeof(indexed_captures_),mem::AbstractDict,x::DupSubpatternNumbers,context,reset_index)

set `reset_index===true'.

deepmap_parser(::typeof(indexed_captures_),mem::AbstractDict,x::Either,context,reset_index)

Method dispatch, resetting lastindex(context.subroutines) if `reset_index===true'.

deepmap_parser(f::typeof(indexed_captures_),mem::AbstractDict,x::Capture,context,a...)

Map the capture my setting index to nextind(context,x).

Registers result in context.subroutines if no previous subroutine with the same index exists (see also DupSubpatternNumbers).

deepmap_parser(f::Function,x::CombinedParser,a...;kw...)

Perform a deep transformation of a CombinedParser. Used for log_names.

Calls deepmap_parser(f,IdDict(),x,a...).

deepmap_parser(f::Function,mem::AbstractDict,x,a...;kw...)

Perform a deep transformation of a CombinedParser.

Note

For a custom parser P<:CombinedParser with sub-parsers, provide a method

deepmap_parser(f::Function,mem::AbstractDict,x::P,a...;kw...) =
    get!(mem,x) do
        ## construct replacement, e.g. if P <: WrappedParser
        P(deepmap_parser(f,mem,x.parser,a...;kw...))
    end

deepmap_parser(f::Function,mem::AbstractDict,x::LeafParser,a...;kw...)

return

    get!(mem,x) do
        f(x,a...;kw...)
    end

CombinedParsers.Regexp.NoDict — Type

For use in ParserWithCaptures to enforce different indices for identical captures.