Internal API

Iterating Parsings

CombinedParsers._iterate — Function

_iterate(p::ParserWithCaptures, sequence::AbstractString)

Base.get — Function

Base.get(parser::CombinedParser{Nothing}, sequence, till, after, i, state)

Default method for parser types returning nothing

Base.get(parser::Transformation, a...)

Constant value parser.transform fallback method.

Base.get(parser::Transformation{<:Function}, a...)

Function call parser.transform(get(parser.parser,a...),i).

Base.get(parser::Transformation{<:IndexAt}, a...)

getindex(get(parser.parser,a...).parser.transform)

CombinedParsers.ismatch — Function

" ismatch(c::Char,p::Union{Function,Steprange,Set,UnicodeClass})::Bool

checks is a character matches a pattern.

Rewriting Parsers

CombinedParsers.deepmap_parser — Function

deepmap_parser(f::typeof(indexed_captures_),mem::AbstractDict,x::Capture,context,a...)

Map the capture my setting index to nextind(context,x).

Registers result in context.subroutines if no previous subroutine with the same index exists (see also DupSubpatternNumbers).

deepmap_parser(f::Function,x::CombinedParser,a...;kw...)

Perform a deep transformation of a CombinedParser. Used for log_names.

Calls deepmap_parser(f,IdDict(),x,a...).

deepmap_parser(f::Function,mem::AbstractDict,x,a...;kw...)

Perform a deep transformation of a CombinedParser.

Note

For a custom parser P<:CombinedParser with sub-parsers, provide a method

deepmap_parser(f::Function,mem::AbstractDict,x::P,a...;kw...) =
    get!(mem,x) do
        ## construct replacement, e.g. if P <: WrappedParser
        P(deepmap_parser(f,mem,x.parser,a...;kw...))
    end

deepmap_parser(f::Function,mem::AbstractDict,x::LeafParser,a...;kw...)

return

    get!(mem,x) do
        f(x,a...;kw...)
    end

Parsing Options

CombinedParsers.Regexp.ParserOptions — Type

A wrapper matching the inner parser on with_options(set_flags, unset_flags, sequence).

CombinedParsers.Regexp.FilterOptions — Type

Lazy wrapper for a sequence, masking elements in getindex with MatchingNever if any of flags are not set.

TODO: make flags a filter function?

CombinedParsers.Regexp.WithOptions — Type

A lazy element transformation type (e.g. AbstractString), getindex wraps elements in with_options(flags,...).

With parsing options

TODO: make flags a transformation function?

CombinedParsers.Regexp.OnOptionsParser — Type

Parser wrapper sequence with if_options.

Internal Types

Abstract Parser Types

CombinedParsers.ParserTypes — Constant

Julia types that can convert(AbstractToken,x). TODO: remove

CombinedParsers.CombinedParser — Type

CombinedParser{T} <: AbstractToken{T}

Abstract parser type for parsers returning matches transformed to ::T.

CombinedParsers.LeafParser — Type

LeafParser{T} <: AbstractToken{T}

Abstract parser type for parsers that have no sub-parser. Used for dispatch in deepmap_parser

CombinedParsers.ConstantParser — Type

wrapper for stepping with ncodeunit length.

CombinedParsers.NIndexParser — Type

Abstract type for stepping with previndex/nextindex, accounting for ncodeunit length of chars at point.

CombinedParsers.WrappedParser — Type

Abstract type for parser wrappers, providing default methods

CombinedParsers.LookAround — Type

Parsers that do not consume any input can inherit this type.

CombinedParsers.Transformation — Type

Transformation{T}(transform::Function, p_) where {T}
map_at(f::Function, p, a...)
map_at(f::Function, Tc::Type, p, a...)
instance(Tc::Type, p::ParserTypes, a...)
instance(Tc::Type, p::ParserTypes)
Base.map(f::Function, Tc::Type, p::ParserTypes, a...)
Base.map(f::Function, p::ParserTypes, a...)

Parser transforming result of a wrapped parser. a... is passed as additional arguments to f (at front .

Regular Expression Types

CombinedParsers.Regexp.ParseMatch — Type

Wrapper type for SequenceWithCaptures, providing getindex and getproperty behavior like RegexMatch.

julia> m = match(re"(?<a>so)+ (or)", "soso or")
ParseMatch("soso or", a="so", 2="or")

julia> m[:a]
"so"

julia> m[2]
"or"

julia> m.match, m.captures
("soso or", SubString{String}["so", "or"])

CombinedParsers.Regexp.Capture — Type

Capture a parser result, optionally with a name. index field is initialized when calling ParserWithCaptures on the parser.

ParserWithCaptures

CombinedParsers.Regexp.Backreference — Type

Backreference(f::Function,index::Integer)

Backreference(f::Function,name::Union{Nothing,Symbol},index::Integer)

Backreference(f::Function,name::AbstractString)

Parser matching previously captured sequence, optionally with a name. index field is recursively set when calling 'ParserWithCaptures` on the parser.

CombinedParsers.Regexp.Subroutine — Type

Parser matching preceding capture, optionally with a name. index field is recursively set when calling ParserWithCaptures on the parser.

CombinedParsers.Regexp.Conditional — Type

Conditional parser, _iterate cycles conditionally on _iterate_condition through matches in field yes and no respectively.

CombinedParsers.Regexp.DupSubpatternNumbers — Type

Parser wrapper for ParserWithCaptures, setting resetindex=true in `deepmapparser(::typeof(indexedcaptures),...)`.

julia> p = re"(?|(a)|(b))\1"
🗄 Sequence |> regular expression combinator with 1 capturing groups
├─ |🗄... Either |> DupSubpatternNumbers
│  ├─ (a)  |> Capture 1
│  └─ (b)  |> Capture 1
└─ \g{1} Backreference
::Tuple{Char,AbstractString}

julia> match(p, "aa")
ParseMatch("aa", 1="a")

julia> match(p, "bb")
ParseMatch("bb", 1="b")

See also pcre doc

CombinedParsers.Regexp.ParserWithCaptures — Type

Top level parser supporting regular expression features captures, backreferences and subroutines. Collects subroutines in field subroutines::Vector and indices of named capture groups in field names::Dict.

Note

implicitly called in match

See also Backreference, Capture, Subroutine

CombinedParsers.Regexp.SequenceWithCaptures — Type

SequenceWithCaptures ensapsulates a sequence to be parsed, and parsed captures.

See also ParserWithCaptures