Public Documentation

Documentation for CombinedParsers.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Public Interface

CombinedParsers

CombinedParsers — Module

A package for combining parsers and transforming strings into julia types.

Compose parsers with the functional parser combinator paradigm, utilize Julia's type inferrence for transformations, log conveniently for debugging, and let Julia compile your parser for good performance.

CombinedParsers.Regexp

CombinedParsers.Regexp — Module

A regular expression parser transforming a PCRE string to a CombinedParser equivalent to the regular expression.

CombinedParsers.Regexp.@re_str — Macro

@re_str(x,flags)

Plug-in replacement for PCRE string macro @r_str. Constructs a CombinedParser equivalent to the PCRE regular expression. Supported flags are parsed

xx Base.PCRE.EXTENDED_MORE

i Base.PCRE.CASELESS

m Base.PCRE.MULTILINE

n Base.PCRE.NOAUTOCAPTURE

U Base.PCRE.UNGREEDY

J Base.PCRE.DUPNAMES

s Base.PCRE.DOTALL

x Base.PCRE.EXTENDED

Unsupported

g UInt32(0),

B UInt32(0)

julia> re"a|b|c"i
|🗄... Either |> regular expression combinator
├─ [aA]
├─ [bB]
└─ [cC]
::Char

julia> re"(a|b)+c"
🗄 Sequence |> regular expression combinator with 1 capturing groups
├─ (|🗄...)+ Either |> Capture |> Repeat
│  ├─ a
│  └─ b
└─ c
::Tuple{Array{Char,1},Char}

julia> re"([ab]+c)*"
(🗄)* Sequence |> Capture |> Repeat |> regular expression combinator with 1 capturing groups
├─ [ab]+  |> Repeat
└─ c
::Array{Tuple{Array{Char,1},Char},1}

Missing docstring.

Missing docstring for CombinedParsers.Regexp.match. Check Documenter's build log for details.

CombinedParsers.regex_escape — Function

regex_escape(s::AbstractString)

regular expression metacharacters are escaped along with whitespace.

Examples

julia> regex_escape("Bang!")
"Bang\!"
julia> regex_escape("  ( [ { . ? *")
"\ \ \(\ \[\ \{\ \.\ \?\ \*"
julia> regex_escape("/^[a-z0-9_-]{3,16}$/")
"/\^\[a\-z0\-9_\-\]\{3,16\}\$/"

Parsing, Logging and Side-Effects

Base.parse — Function

parse(parser::ParserTypes, str::AbstractString)

Parse a string with a CombinedParser as an instance of result_type(parser).

CombinedParsers.with_name — Function

with_name(name::Symbol,x; doc="")

A parser labelled with name. Labels are useful in printing and logging.

See also: @with_names, with_name, log_names

CombinedParsers.log_names — Function

log_names(x,names=nothing; exclude=nothing)

Rebuild parser replacing NamedParser instances with with_log parsers.

See also: with_log

CombinedParsers.with_log — Function

with_log(s::AbstractString,p, delta=5;nomatch=false)

Log matching process of parser p, displaying delta characters left of and right of match.

If nomatch==true, also log when parser does not match.

See also: log_names, with_effect

Missing docstring.

Missing docstring for instrument. Check Documenter's build log for details.

CombinedParsers.with_effect — Function

with_effect(f::Function,p,a...)

Call f(sequence,before_i,after_i,state,a...) if p matches, f(sequence,before_i,before_i,nothing,a...) otherwise.

Transformations

Base.map — Function

map(f::Function, p::TextParse.AbstractToken, a...)

Parser matching p, transforming parsing results (x) with function f(x,a...).

See also: map_at

map(T::Type, p::TextParse.AbstractToken, a...)

Parser matching p, transforming parsing results (x) with constructor T(x,a...).

See also: map_at

CombinedParsers.map_at — Function

map_at(f::Function, p, a...)
map_at(f::Function, Tc::Type, p, a...)

Parser transforming result of a wrapped parser. a... is passed as additional arguments to f.

CombinedParsers.JoinSubstring — Type

JoinSubstring(x)
(!)(x::TextParse.AbstractToken)

Parser Transformation getting the matched SubString.

Parser Constructors

Character Matchers

CombinedParsers.AnyChar — Type

AnyChar()

Parser matching exactly one Char, returning the character.

julia> AnyChar()
re"."

CombinedParsers.CharIn — Type

CharIn(x)

Parser matching exactly one character in x.

julia> a_z = CharIn('a':'z')
re"[a-z]"

julia> parse(a_z, "a")
'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)

julia> ac = CharIn("ac")
re"[ac]"

julia> parse(ac, "c")
'c': ASCII/Unicode U+0063 (category Ll: Letter, lowercase)

CombinedParsers.CharNotIn — Type

CharNotIn(x)

Parser matching exactly one character in x.

julia> a_z = CharNotIn('a':'z')
re"[^a-z]"

julia> parse(a_z, "a")
ERROR: ArgumentError: no successfull parsing.
Stacktrace:
 [1] parse(::CharNotIn{Tuple{StepRange{Char,Int64}}}, ::String) at /home/gregor/dev/julia/CombinedParsers/src/CombinedParsers.jl:2583
 [2] top-level scope at REPL[24]:1

julia> ac = CharNotIn("ac")
re"[^ac]"

julia> parse(ac, "c")
ERROR: ArgumentError: no successfull parsing.
Stacktrace:
 [1] parse(::CharNotIn{Array{Char,1}}, ::String) at /home/gregor/dev/julia/CombinedParsers/src/CombinedParsers.jl:2583
 [2] top-level scope at REPL[24]:1

Combining Parser

CombinedParsers.Either — Type

Either(parsers...)

Parser that tries matching the provided parsers in order, accepting the first match, and fails if all parsers fail.

julia> match(r"a|bc","bc")
RegexMatch("bc")

julia> parse(Either("a","bc"),"bc")
"bc"

julia> parse("a" | "bc","bc")
"bc"

Missing docstring.

Missing docstring for Base.(|). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Base.:| :: Tuple{Either,Type}. Check Documenter's build log for details.

CombinedParsers.sEither — Function

sEither(x...)

Simplifying Either, flattens nested Eithers, remove Never parsers.

julia> Either('a',CharIn("AB")|"bc")
|🗄... Either
├─ a
└─ |🗄... Either
   ├─ [AB]
   └─ bc
::Union{Char, SubString}

julia> sEither('a',CharIn("AB")|"bc")
|🗄... Either
├─ a
├─ [AB]
└─ bc
::Union{Char, SubString}

See also Either

CombinedParsers.Sequence — Type

Sequence(;kw...)

Sequence keyword argument constructors transform the parsing into a named tuple.

CombinedParsers.sSequence — Function

sSequence(x...)

Simplifying Sequence, flatten Sequences, remove Always lookarounds.

julia> Sequence('a',CharIn("AB")*'b')
🗄 Sequence
├─ a
└─ 🗄 Sequence
   ├─ [AB]
   └─ b
::Tuple{Char,Tuple{Char,Char}}

julia> sSequence('a',CharIn("AB")*'b')
🗄 Sequence
├─ a
├─ [AB]
└─ b
::Tuple{Char,Char,Char}

See also Sequence

Missing docstring.

Missing docstring for FlatMap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for after. Check Documenter's build log for details.

CombinedParsers.Atomic — Type

Atomic(x)

A parser matching p, and failing when required to backtrack (behaving like an atomic group in regular expressions).

Repeating

CombinedParsers.Optional — Type

Optional(parser;default=defaultvalue(result_type(parser)))

Parser that always succeeds. If parser succeeds, return result of parser with curser behind match. If parser does not succeed, return default with curser unchanged.

julia> match(r"a?","b")
RegexMatch("")

julia> parse(Optional("a"),"b")
""

Missing docstring.

Missing docstring for Base.(|)(x::TextParse.AbstractToken{T}, default::Union{T,Missing}) where { T }. Check Documenter's build log for details.

CombinedParsers.Repeat — Type

Repeat(x)
Repeat(x; min=0,max=typemax(Int))
Repeat(min::Integer, x)
Repeat(min::Integer,max::Integer, x)

Parser repeating pattern x min:max times.

CombinedParsers.Repeat1 — Function

Repeat1(x)

Parser repeating pattern x one time or more.

Repeat1(f::Function,a...)

Abbreviation for map(f,Repeat1(a...)).

CombinedParsers.Repeat_stop — Function

Repeat_stop(p,stop)

Repeat p until stop (NegativeLookahead), not matching stop. Sets cursor before stop. Returns results of p.

julia> p = Repeat_stop(AnyChar(),'b') * AnyChar()
🗄 Sequence
├─ 🗄* Sequence |> map(#41) |> Repeat
│  ├─ (?!🗄)
│  │  └─ b
│  └─ .
└─ .
::Tuple{Array{Char,1},Char}

julia> parse(p,"acbX")
(['a', 'c'], 'b')

See also NegativeLookahead

CombinedParsers.Repeat_until — Function

Repeat\_until(p,until, with_until=false;wrap=identity)

Repeat p until stop (with Repeat_stop). and set point after stop.

Return a Vector{result_type(p)} if wrap_until==false, otherwise a Tuple{Vector{result_type(p)},result_type(until)}.

To transform the Repeat_stop(p) parser head, provide a function(::Vector{result_type(p)}) in wrap keyword argument, e.g.

julia> p = Repeat_until(AnyChar(),'b') * AnyChar()
🗄 Sequence
├─ 🗄 Sequence |> map(#41)
│  ├─ 🗄* Sequence |> map(#41) |> Repeat
│  │  ├─ (?!🗄)
│  │  │  └─ b
│  │  └─ .
│  └─ b
└─ .
::Tuple{Array{Char,1},Char}

julia> parse(p,"acbX")
(['a', 'c'], 'X')

julia> parse(Repeat_until(AnyChar(),'b';wrap=JoinSubstring),"acbX")
"ac"

See also NegativeLookahead

Base.join — Function

Base.join(x::Repeat,delim)

Parser matching repeated x.parser separated by delim.

julia> parse(join(Repeat(AnyChar()),','),"a,b,c")
3-element Array{Char,1}:
 'a'
 'b'
 'c'

Assertions

CombinedParsers.Always — Type

Always()

Assertion parser matching always and not consuming any input. Returns Always().

julia> Always()
re""

CombinedParsers.Never — Type

Never()

Assertion parser matching never and not consuming any input.

julia> Never()
re"(*FAIL)"

CombinedParsers.PositiveLookahead — Type

PositiveLookahead(parser)

Parser that succeeds if and only if parser succeeds, but consumes no input. The match is returned. Useful for checks like "must be followed by parser, but don't consume its match".

julia> la=PositiveLookahead("peek")
re"(?=peek)"

julia> parse(la*AnyChar(),"peek")
(re"(?=peek)", 'p')

CombinedParsers.NegativeLookahead — Type

NegativeLookahead(parser)

Parser that succeeds if and only if parser does not succeed, but consumes no input. parser is returned as match. Useful for checks like "must not be followed by parser, don't consume its match".

julia> la = NegativeLookahead("peek")
re"(?!peek)"

julia> parse(la*AnyChar(),"peek")
ERROR: ArgumentError: no successfull parsing.
Stacktrace:
 [1] parse(::Sequence{Tuple{SubString,Char},Tuple{NegativeLookahead{SubString,CombinedParsers.ConstantParser{4,SubString}},AnyChar}}, ::String) at /home/gregor/dev/julia/CombinedParsers/src/CombinedParsers.jl:2583
 [2] top-level scope at REPL[24]:1

CombinedParsers.PositiveLookbehind — Type

wraps a parser::P, succeeds if and only if parser succeeds, but consumes no input. The match is returned. Useful for checks like "must be followed by parser, but don't consume its match".

CombinedParsers.NegativeLookbehind — Type

wraps a parser::P, succeeds if and only if parser does not succeed, but consumes no input. nothing is returned as match. Useful for checks like "must not be followed by parser, don't consume its match".

Printing

CombinedParsers.regex_string — Function

regex_string(x::AbstractParser)

regex_prefix(x)*regex_inner(x)*regex_suffix(x)

CombinedParsers.regex_prefix — Function

regex_prefix(x)

Prefix printed in parser tree node.

CombinedParsers.regex_inner — Function

regex_inner(x::TextParse.AbstractToken)

Regex representation of x. See regex_string

CombinedParsers.regex_suffix — Function

regex_suffix(x)

Suffix printed in parser tree node.

Internals

CombinedParsers.AbstractParser — Type

AbstractParser{T} <: TextParse.AbstractToken{T}

Abstract parser type for parsers returning matches transformed to ::T.

CombinedParsers.ConstantParser — Type

wrapper for stepping with ncodeunit length.

CombinedParsers.NIndexParser — Type

Abstract type for stepping with previndex/nextindex, accounting for ncodeunit length of chars at point.

CombinedParsers.WrappedParser — Type

Abstract type for parser wrappers, providing default methods

CombinedParsers.LookAround — Type

Parsers that do not consume any input can inherit this type

CombinedParsers._iterate — Function

parse(p::ParserTypes, s::AbstractString)

parse s with parser p.

_iterate(parser, sequence, till, i, states)

Note: i is the index in sequence after parser match according to state (and not the start of the match), such that start_index(sequence,after,parser,state) returns the start of the matching subsequence, and sequence[start_index(sequence,after,parser,state):prevind(sequence,i)] is the matched subsequence.