Public Documentation

Documentation for CombinedParsers.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Matching

CombinedParsers — Module

A package for combining parsers and transforming strings into julia types.

Compose parsers with the functional parser combinator paradigm, utilize Julia's type inferrence for transformations, log conveniently for debugging, and let Julia compile your parser for good performance.

Base.match — Function

Base.match(parser::ParserTypes,sequence::AbstractString[, idx::Integer]; log=nothing)

Plug-in replacement for match(::Regex,sequence).

CombinedParsers.MatchesIterator — Type

Iterator type for match_all and parse_all.

CombinedParsers.ParseMatch — Type

Wrapper type for CombinedParsers.Regexp.SequenceWithCaptures, providing getindex and getproperty behavior like RegexMatch.

julia> m = match(re"(?<a>so)+ (or)", "soso or")
ParseMatch("soso or", a="so", 2="or")

julia> m[:a]
"so"

julia> m[2]
"or"

julia> m.match, m.captures
("soso or", SubString{String}["so", "or"])

Base.:== — Method

==(pcre_m::RegexMatch,pc_m::ParseMatch)

equal iif values of .match, .offset, .ncodeunits and .captures are equal.

Base.getindex — Method

Base.getindex(x::ParseMatch{<:Any,<:SequenceWithCaptures,<:Any},i::Union{Integer,Symbol})

Gets capture i as SubString.

See API of RegexMatch.

Base.getproperty — Method

Base.getproperty(m::ParseMatch{<:Any,<:SequenceWithCaptures,<:Any},key::Symbol)

enable m.captures and m.match.

See API of RegexMatch.

Parsing

Base.get — Function

Base.get(x::ParseMatch{<:MatchTuple})

Get the result of a match result.

```jldoctest julia> m = match(re"(?<a>so)+ (or)", "soso or") ParseMatch("soso or", a="so", 2="or")

julia> get(m) "so"

julia> m[2] "or"

julia> m.match, m.captures ("soso or", SubString{String}["so", "or"])

```

Base.get(parser::AbstractToken{Nothing}, sequence, till, after, i, state)

Default method for parser types returning nothing

Base.get(parser::AbstractToken{Nothing}, sequence, till, after, i, state)

Default method for parser types returning nothing

Base.get(parser::Transformation, a...)

Constant value parser.transform fallback method.

Base.get(parser::Transformation{<:Constant}, a...)

does not evaluate get(parser.transform,...).

Base.get(parser::Transformation{<:Function}, a...)

Function call parser.transform(get(parser.parser,a...)).

Base.get(parser::Transformation{<:Type}, a...)

Function call parser.transform(get(parser.parser,a...)).

Base.get(parser::Transformation{<:IndexAt}, a...)

getindex(get(parser.parser,a...).parser.transform)

Base.parse — Function

parse(parser::ParserTypes, str::AbstractString; log=nothing)

Parse a string with a CombinedParser as an instance of result_type(parser).

If log is a Vector{Symbol}, parser is transformed with log_names(p, log). See also log_names.

DocTestFilters = r"map\(.+\)"

julia> using TextParse

julia> p = ("Number: "*TextParse.Numeric(Int))[2]
🗄 Sequence |> map(IndexAt(2))
├─ Number\:\
└─ <Int64>
::Int64

julia> parse(p,"Number: 42")
42

Base.tryparse — Function

tryparse(parser::ParserTypes, str::AbstractString[, idx=1])

Like parse, but returns either a value of result_type(parser) or nothing if string does not start with with a match.

CombinedParsers.tryparse_pos — Function

tryparse_pos(parser::ParserTypes, str::AbstractString)

Like parse, but returns either a tuple of result_type(parser) and the position after the match, or nothing if string does not start with with a match.

Typed Transformation Parsers

CombinedParsers.state_type — Function

CombinedParsers.state_type(x::T)

Return the state type of x

Base.map — Function

map(f::Function, p::AbstractToken, a...)

Parser matching p, transforming parsing results (x) with function f(x,a...).

See also: map_at, Transformation

map(T::Type, p::AbstractToken, a...)

Parser matching p, transforming ps parsing result with constructor T(x,a...).

See also: map_at get, Transformation

map(index::IndexAt, p::AbstractToken, a...)
map(constant, p::AbstractToken, a...)

Parser matching p, transforming ps parsing results to getindex(x,index) or constant.

See also: map_at get, Transformation

Missing docstring.

Missing docstring for map_at. Check Documenter's build log for details.

CombinedParsers.JoinSubstring — Type

JoinSubstring(x)
(!)(x::AbstractToken)

Parser Transformation getting the matched SubString.

CombinedParsers.Constant — Type

A Transformation{<:Constant} skips evaluation of get(.parser).

Base.:! — Method

(!)(x::AbstractToken)

Parser Transformation getting the matched SubString.

julia> parse(Repeat(CharIn(:L)),"abc123")
3-element Array{Char,1}:
 'a'
 'b'
 'c'

julia> parse(!Repeat(CharIn(:L)),"abc123")
"abc"

CombinedParsers.result_type — Function

result_type(::Type{T}) where {T<:Tuple}

Internally used for Sequence result_type.

Parser Constructors

CombinedParsers.parser — Function

parser(x::Union{AbstractString,Char})

A ConstantParser matching x.

parser(x::StepRange{Char,<:Integer})

CharIn matching x.

parser(x::Pair{Symbol, P}) where P

A parser labelled with name x.first. Labels are useful in printing and logging.

See also: @with_names, with_name, log_names

parser(constant::Pair{<:ParserTypes})
map_constant(constant, parser::AbstractToken)

A parser mapping matches of parser x.first to constant x.second.

Base.convert — Function

convert(::Type{AbstractToken},x::Union{AbstractString,Char})

A ConstantParser matching x.

CombinedParsers.Regexp.@re_str — Macro

@re_str(x,flags)

Construct a ParserWithCaptures from PCRE regex syntax, such as re"^[a-z]*$", without interpolation and unescaping (except for quotation mark " which still has to be escaped). Plug-in replacement for PCRE string macro @r_str.

The regex also accepts one or more flags, listed after the ending quote, to change its behaviour:

i enables case-insensitive matching

m treats the ^ and $ tokens as matching the start and end of individual lines, as opposed to the whole string.

s allows the . modifier to match newlines.

x enables "comment mode": whitespace is ignored except when escaped with \, and # is treated as starting a comment.

a disables UCP mode (enables ASCII mode). By default \B, \b, \D, \d, \S, \s, \W, \w, etc. match based on Unicode character properties. With this option, these sequences only match ASCII characters.

xx enables "extended comment mode": whitespace in bracket character matchers are ignored.

julia> re"a|c"i
|🗄... Either
├─ [Aa] CharIn
└─ [cC] CharIn
::Char

julia> re"a+c"
🗄 Sequence
├─ a+  |> Repeat
└─ c
::Tuple{Array{Char,1},Char}

See also Regcomb, parse_options.

useful importable sub-parsers

julia> using CombinedParsers.Regexp

julia> import CombinedParsers.Regexp.whitespace_horizontal

julia> import CombinedParsers.Regexp.whitespace_maybe

julia> import CombinedParsers.Regexp.word

CombinedParsers.Regexp.integer_base — Function

integer_base(base,mind=0,maxd=Repeat_max)

Parser matching a integer format on base base.

Character Matchers

CombinedParsers.AnyChar — Type

AnyChar()

Parser matching exactly one Char, returning the character.

julia> AnyChar()
re"."

CombinedParsers.CharIn — Type

CharIn(x)

Parser matching exactly one character in x.

julia> a_z = CharIn('a':'z')
re"[a-z]"

julia> parse(a_z, "a")
'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)

julia> ac = CharIn("ac")
re"[ac]"

julia> parse(ac, "c")
'c': ASCII/Unicode U+0063 (category Ll: Letter, lowercase)

CombinedParsers.CharNotIn — Type

CharNotIn(x)

Parser matching exactly one character in x.

julia> a_z = CharNotIn('a':'z')
re"[^a-z]"

julia> ac = CharNotIn("ac")
re"[^ac]"

Combining Parser

CombinedParsers.@syntax — Macro

Convenience macro @syntax name = expr for defining a CombinedParser name=expr and custom parsing macro @name_str.

With @syntax for name in either; expr; end the defined parser is pushfirst! to either. If either is undefined, it will be created. If either == :text || either == Symbol(:) the parser will be added to CombinedParser_globals variable in your module.

DocTestFilters = r"map\(#[^)]+\)"

julia> @syntax for german_street_address in street_address
            Sequence(:street => !Repeat(AnyChar()),
                 " ",
                 :no =>Numeric(Int))
       end
🗄 Sequence |> map(#38) |> with_name(:german_street_address)
├─ .* AnyChar |> Repeat |> JoinSubstring |> with_name(:street)
├─ \
└─ TextParse.Numeric{Int64} TextParse.Numeric |> with_name(:no)
::NamedTuple{(:street, :no),Tuple{SubString,Int64}}

julia> german_street_address"Some Avenue 42"
NamedTuple{(:street, :no),Tuple{SubString,Int64}}(("Some Avenue", 42))

julia> @syntax for us_street_address in street_address
            Sequence(:no =>Numeric(Int),
                     " ",
                     :street => !Repeat(AnyChar()))
       end
🗄 Sequence |> map(#38) |> with_name(:us_street_address)
├─ TextParse.Numeric{Int64} TextParse.Numeric |> with_name(:no)
├─ \  
└─ .* AnyChar |> Repeat |> JoinSubstring |> with_name(:street)
::NamedTuple{(:no, :street),Tuple{Int64,SubString}}

julia> street_address"50 Oakland Ave"
NamedTuple{(:no, :street),Tuple{Int64,SubString}}((50, "Oakland Ave"))

CombinedParsers.Either — Type

Parser that tries matching the provided parsers in order, accepting the first match, and fails if all parsers fail.

This parser has no custom == and hash methods because it can recurse.

julia> match(r"a|bc","bc")
RegexMatch("bc")

julia> parse(Either("a","bc"),"bc")
"bc"

julia> parse("a" | "bc","bc")
"bc"

Missing docstring.

Missing docstring for CombinedParsers.either_types. Check Documenter's build log for details.

Base.:| — Method

(|)(x::AbstractToken, y)
(|)(x, y::AbstractToken)
(|)(x::AbstractToken, y::AbstractToken)

Operator syntax for sEither(x, y).

julia> 'a' | CharIn("AB") | "bc"
|🗄... Either
├─ a
├─ [AB] CharIn
└─ bc
::Union{Char, SubString}

Base.push! — Function

Base.push!(x::Either, option)

Push option to x.options as parser tried next if x fails. Recursive parsers can be built with push! to Either.

See also pushfirst!.

Base.pushfirst! — Function

Base.pushfirst!(x::Either, option)

Push option to x.options as parser tried first, and trying x if option fails. Recursive parsers can be built with pushfirst! to Either.

See also push!.

CombinedParsers.sEither — Function

sEither(x...)

Simplifying Either, flattens nested Eithers, remove Never parsers.

julia> Either('a',CharIn("AB")|"bc")
|🗄... Either
├─ a
└─ |🗄... Either
   ├─ [AB] CharIn
   └─ bc
::Union{Char, SubString}

julia> sEither('a',CharIn("AB")|"bc")
|🗄... Either
├─ a
├─ [AB] CharIn
└─ bc
::Union{Char, SubString}

See also Either

CombinedParsers.Sequence — Type

Sequence(;kw...)

Sequence keyword argument constructors transform the parsing into a named tuple.

CombinedParsers.sSequence — Function

sSequence(x...)

Simplifying Sequence, flatten Sequences, remove Always lookarounds.

julia> Sequence('a',CharIn("AB")*'b')
🗄 Sequence
├─ a
└─ 🗄 Sequence
   ├─ [AB] CharIn
   └─ b
::Tuple{Char,Tuple{Char,Char}}

julia> sSequence('a',CharIn("AB")*'b')
🗄 Sequence
├─ a
├─ [AB] CharIn
└─ b
::Tuple{Char,Char,Char}

See also Sequence

CombinedParsers.Atomic — Type

Atomic(x)

A parser matching p, and failing when required to backtrack (behaving like an atomic group in regular expressions).

Repeating

CombinedParsers.Optional — Type

Optional(parser;default=defaultvalue(result_type(parser)))

Parser that always succeeds. If parser succeeds, return result of parser with curser behind match. If parser does not succeed, return default with curser unchanged.

julia> match(r"a?","b")
RegexMatch("")

julia> parse(Optional("a", default=42),"b")
42

Base.:| — Method

`(|)(x::Either, T::Type)`

Return new Either with T added to result_type(x). todo: Note that the options array is kept. As a consequence push!on result will also push to x.

Base.:| — Method

(|)(x::AbstractToken{T}, default::Union{T,Missing})

Operator syntax for Optional(x, default=default).

julia> parser("abc") | "nothing"
|🗄... Either
├─ abc
└─ nothing
::SubString

CombinedParsers.Repeat — Type

Repeat(x)
Repeat(x; min=0,max=Repeat_max)
Repeat(min::Integer, x)
Repeat(min::Integer,max::Integer, x)

Parser repeating pattern x min:max times.

CombinedParsers.Lazy — Type

Lazy(x::Repeat)
Lazy(x::Optional)

Lazy x repetition matching from greedy to lazy.

julia> re"a+?"
a+?  |> Repeat |> Lazy
::Array{Char,1}

julia> re"a??"
a??  |> Optional(default=missing) |> Lazy
::Union{Missing, Char}

CombinedParsers.Repeat1 — Function

Repeat1(x)

Parser repeating pattern x one time or more.

Repeat1(f::Function,a...)

Abbreviation for map(f,Repeat1(a...)).

CombinedParsers.Repeat_stop — Function

Repeat_stop(p,stop)

Repeat p until stop (NegativeLookahead), not matching stop. Sets cursor before stop. Returns results of p.

julia> p = Repeat_stop(AnyChar(),'b') * AnyChar()
🗄 Sequence
├─ 🗄* Sequence |> map(#52) |> Repeat
│  ├─ (?!🗄) NegativeLookahead
│  │  └─ b
│  └─ . AnyChar
└─ . AnyChar
::Tuple{Array{Char,1},Char}

julia> parse(p,"acbX")
(['a', 'c'], 'b')

See also NegativeLookahead

CombinedParsers.Repeat_until — Function

Repeat\_until(p,until, with_until=false;wrap=identity)

Repeat p until stop (with Repeat_stop). and set point after stop.

Return a Vector{result_type(p)} if wrap_until==false, otherwise a Tuple{Vector{result_type(p)},result_type(until)}.

To transform the Repeat_stop(p) parser head, provide a function(::Vector{result_type(p)}) in wrap keyword argument, e.g.

julia> p = Repeat_until(AnyChar(),'b') * AnyChar()
🗄 Sequence
├─ 🗄 Sequence |> map(#52)
│  ├─ (?>🗄*) Sequence |> map(#52) |> Repeat |> Atomic
│  │  ├─ (?!🗄) NegativeLookahead
│  │  │  └─ b
│  │  └─ . AnyChar
│  └─ b
└─ . AnyChar
::Tuple{Array{Char,1},Char}

julia> parse(p,"acbX")
(['a', 'c'], 'X')

julia> parse(Repeat_until(AnyChar(),'b';wrap=JoinSubstring),"acbX")
"ac"

See also NegativeLookahead

Base.join — Function

Base.join(x::Repeat,delim, infix=:skip)

Parser matching repeated x.parser separated by delim.

julia> parse(join(Repeat(AnyChar()),','),"a,b,c")
3-element Array{Char,1}:
 'a'
 'b'
 'c'

julia> parse(join(Repeat(AnyChar()),',';infix=:prefix),"a,b,c")
('a', [(',', 'b'), (',', 'c')])

julia> parse(join(Repeat(AnyChar()),',';infix=:suffix),"a,b,c")
([('a', ','), ('b', ',')], 'c')

Base.join(x::Repeat,delim)

Shorthand for join(Repeat(x),delim).

Base.join(f::Function, x::CombinedParser, delim)

Shorthand for map(f,join(x,delim)).

Logging and Side-Effects

CombinedParsers.with_name — Function

with_name(name::Symbol,x; doc="")

A parser labelled with name. Labels are useful in printing and logging.

See also: @with_names, with_name, log_names

CombinedParsers.@with_names — Macro

Sets names of parsers within begin/end block to match the variables they are asigned to.

so, for example

julia> @with_names foo = AnyChar()
. AnyChar |> with_name(:foo)
::Char

julia> parse(log_names(foo),"ab")
   match foo: ab
              ^
'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)

See also log_names(parser)

CombinedParsers.log_names — Function

log_names(x,names=true; exclude=nothing)

Rebuild parser replacing NamedParser instances with with_log parsers. Log all NamedParser instanses if names==true or name in names and not name in exclude.

See also: with_log, deepmap_parser

CombinedParsers.with_log — Function

with_log(s::AbstractString,p, delta=5;nomatch=false)

Log matching process of parser p, displaying delta characters left of and right of match.

If nomatch==true, also log when parser does not match.

See also: log_names, with_effect

CombinedParsers.with_effect — Function

with_effect(f::Function,p,a...)

Call f(sequence,before_i,after_i,state,a...) if p matches, f(sequence,before_i,before_i,nothing,a...) otherwise.

Assertions

CombinedParsers.AtStart — Type

AtStart()

Parser succeding if and only if at index 1 with result_type AtStart.

julia> AtStart()
re"^"

CombinedParsers.AtEnd — Type

AtEnd()

Parser succeding if and only if at last index with result_type AtEnd.

julia> AtEnd()
re"$"

CombinedParsers.Always — Type

Always()

Assertion parser matching always and not consuming any input. Returns Always().

julia> Always()
re""

CombinedParsers.Never — Type

Never()

Assertion parser matching never.

julia> Never()
re"(*FAIL)"

CombinedParsers.Lookahead — Function

Lookahead(does_match::Bool, p)

PositiveLookahead if does_match==true, NegativeLookahead otherwise.

CombinedParsers.PositiveLookahead — Type

PositiveLookahead(parser)

Parser that succeeds if and only if parser succeeds, but consumes no input. The match is returned. Useful for checks like "must be followed by parser, but don't consume its match".

julia> la=PositiveLookahead("peek")
re"(?=peek)"

julia> parse(la*AnyChar(),"peek")
("peek", 'p')

CombinedParsers.NegativeLookahead — Type

NegativeLookahead(parser)

Parser that succeeds if and only if parser does not succeed, but consumes no input. parser is returned as match. Useful for checks like "must not be followed by parser, don't consume its match".

julia> la = NegativeLookahead("peek")
re"(?!peek)"

julia> parse(la*AnyChar(),"seek")
(re"(?!peek)", 's')

CombinedParsers.Lookbehind — Function

Lookbehind(does_match::Bool, p)

PositiveLookbehind if does_match==true, NegativeLookbehind otherwise.

CombinedParsers.PositiveLookbehind — Type

PositiveLookbehind(parser)

Parser that succeeds if and only if parser succeeds before cursor. Consumes no input. The match is returned. Useful for checks like "must be preceded by parser, don't consume its match".

CombinedParsers.NegativeLookbehind — Type

NegativeLookbehind(parser)

Parser that succeeds if and only if parser does not succeed before cursor. Consumes no input. nothing is returned as match. Useful for checks like "must not be preceded by parser, don't consume its match".

julia> la=NegativeLookbehind("keep")
re"(?<!keep)"

julia> parse("peek"*la,"peek")
("peek", re"(?<!keep)")