Public Documentation

Documentation for CombinedParsers.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

CombinedParsers

CombinedParsers — Module

A package for combining parsers and transforming strings into julia types.

Compose parsers with the functional parser combinator paradigm, utilize Julia's type inferrence for transformations, log conveniently for debugging, and let Julia compile your parser for good performance.

CombinedParsers.parser — Function

calls convert(AbstractToken,x).

Base.convert — Function

convert(::Type{AbstractToken},x::Union{AbstractString,Char})

A ConstantParser matching x.

convert(::Type{AbstractToken},x::StepRange{Char,<:Integer})

CharIn matching x.

convert(::Type{AbstractToken},x::Pair{Symbol, P}) where P

A parser labelled with name x.first. Labels are useful in printing and logging.

See also: @with_names, with_name, log_names

convert(::Type{AbstractToken},constant::Pair{<:ParserTypes})

A parser mapping matches of x.first to constant x.second.

See also: map, map_at

convert(::Type{AbstractToken},x::Union{AbstractString,Char})

A ConstantParser matching x.

CombinedParsers.Regexp

CombinedParsers.Regexp — Module

A regular expression parser transforming a PCRE string to a CombinedParser equivalent to the regular expression.

CombinedParsers.Regexp.@re_str — Macro

@re_str(x,flags)

Construct a ParserWithCaptures from PCRE regex syntax, such as re"^[a-z]*$", without interpolation and unescaping (except for quotation mark " which still has to be escaped). Plug-in replacement for PCRE string macro @r_str.

The regex also accepts one or more flags, listed after the ending quote, to change its behaviour:

i enables case-insensitive matching

m treats the ^ and $ tokens as matching the start and end of individual lines, as opposed to the whole string.

s allows the . modifier to match newlines.

x enables "comment mode": whitespace is ignored except when escaped with \, and # is treated as starting a comment.

a disables UCP mode (enables ASCII mode). By default \B, \b, \D, \d, \S, \s, \W, \w, etc. match based on Unicode character properties. With this option, these sequences only match ASCII characters.

xx enables "extended comment mode": whitespace in bracket character matchers are ignored.

julia> re"a|c"i
|🗄... Either |> regular expression combinator
├─ [aA] CharIn
└─ [cC] CharIn
::Char

julia> re"a+c"
🗄 Sequence |> regular expression combinator
├─ a+  |> Repeat
└─ c
::Tuple{Array{Char,1},Char}

Base.match — Function

Base.match(parser::ParserTypes,sequence::AbstractString; log=false)

Plug-in replacement for match(::Regex,sequence).

CombinedParsers.Regexp.with_options — Function

with_options(options::AbstractString,str::AbstractString)

CombinedParsers.Regexp.parse_options — Function

parse_options(options::AbstractString)

Return PCRE option mask parsed from options.

julia> CombinedParsers.Regexp.pcre_options_parser
🗄 Sequence |> map(#52)
├─ ^ AtStart
├─ 🗄+? Sequence |> map(#52) |> Repeat |> map(#44) |> Optional(default=0)
│  ├─ |🗄... Either
│  │  ├─ dupnames  |> map(Constant(0x00000040)) |> with_name(:DUPNAMES)
│  │  ├─ xx  |> map(Constant(0x01000000)) |> with_name(:EXTENDED_MORE)
│  │  ├─ i  |> map(Constant(0x00000008)) |> with_name(:CASELESS)
│  │  ├─ m  |> map(Constant(0x00000400)) |> with_name(:MULTILINE)
│  │  ├─ n  |> map(Constant(0x00002000)) |> with_name(:NO_AUTO_CAPTURE)
│  │  ├─ U  |> map(Constant(0x00040000)) |> with_name(:UNGREEDY)
│  │  ├─ J  |> map(Constant(0x00000040)) |> with_name(:DUPNAMES)
│  │  ├─ s  |> map(Constant(0x00000020)) |> with_name(:DOTALL)
│  │  ├─ x  |> map(Constant(0x00000080)) |> with_name(:EXTENDED)
│  │  ├─ B  |> map(Constant(0x00000000)) |> with_name(:BINCODE)
│  │  └─ I  |> map(Constant(0x00000000)) |> with_name(:INFO)
│  └─ ,?  |> Optional(default=missing)
└─ $ AtEnd
::UInt32

CombinedParsers.regex_escape — Function

regex_escape(s::AbstractString)

regular expression metacharacters are escaped along with whitespace.

Examples

julia> regex_escape("Bang!")
"Bang\!"
julia> regex_escape("  ( [ { . ? *")
"\ \ \(\ \[\ \{\ \.\ \?\ \*"
julia> regex_escape("/^[a-z0-9_-]{3,16}$/")
"/\^\[a\-z0\-9_\-\]\{3,16\}\$/"

Parsing, Logging and Side-Effects

Base.parse — Function

parse(parser::ParserTypes, str::AbstractString; log=nothing)

Parse a string with a CombinedParser as an instance of result_type(parser).

If log is a Vector{Symbol}, parser is transformed with log_names(p, log). See also log_names.

julia> using TextParse

julia> p = ("Number: "*TextParse.Numeric(Int))[2]
🗄 Sequence |> map(#31)
├─ Number\:\
└─ <Int64>
::Int64

julia> parse(p,"Number: 42")
42

Base.tryparse — Function

tryparse(parser::ParserTypes, str::AbstractString)

Like parse, but returns either a value of result_type(parser) or nothing if string does not start with with a match.

CombinedParsers.with_name — Function

with_name(name::Symbol,x; doc="")

A parser labelled with name. Labels are useful in printing and logging.

See also: @with_names, with_name, log_names

CombinedParsers.@with_names — Macro

Sets names of parsers within begin/end block to match the variables they are asigned to.

so, for example

julia> @with_names foo = AnyChar()
. AnyChar |> with_name(:foo)
::Char

julia> parse(log_names(foo),"ab")
   match foo: ab
              ^
'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)

See also log_names(parser)

CombinedParsers.log_names — Function

log_names(x,names=true; exclude=nothing)

Rebuild parser replacing NamedParser instances with with_log parsers. Log all NamedParser instanses if names==true or name in names and not name in exclude.

See also: with_log, deepmap_parser

CombinedParsers.with_log — Function

with_log(s::AbstractString,p, delta=5;nomatch=false)

Log matching process of parser p, displaying delta characters left of and right of match.

If nomatch==true, also log when parser does not match.

See also: log_names, with_effect

Missing docstring.

Missing docstring for instrument. Check Documenter's build log for details.

CombinedParsers.with_effect — Function

with_effect(f::Function,p,a...)

Call f(sequence,before_i,after_i,state,a...) if p matches, f(sequence,before_i,before_i,nothing,a...) otherwise.

Transformations

Base.map — Function

map(f::Function, p::AbstractToken, a...)

Parser matching p, transforming parsing results (x) with function f(x,a...).

See also: map_at, Transformation

map(T::Type, p::AbstractToken, a...)

Parser matching p, transforming ps parsing result with constructor T(x,a...).

See also: map_at get, Transformation

map(index::IndexAt, p::AbstractToken, a...)
map(constant, p::AbstractToken, a...)

Parser matching p, transforming ps parsing results to getindex(x,index) or constant.

See also: map_at get, Transformation

CombinedParsers.map_at — Function

map_at(f::Function, p, a...)
map_at(f::Function, Tc::Type, p, a...)

Parser transforming result of a wrapped parser. a... is passed as additional arguments to f.

See also: map, Transformation

CombinedParsers.JoinSubstring — Type

JoinSubstring(x)
(!)(x::AbstractToken)

Parser Transformation getting the matched SubString.

Base.:! — Method

(!)(x::AbstractToken)

Parser Transformation getting the matched SubString.

julia> parse(Repeat(CharIn(:L)),"abc123")
3-element Array{Char,1}:
 'a'
 'b'
 'c'

julia> parse(!Repeat(CharIn(:L)),"abc123")
"abc"

Parser Constructors

Character Matchers

CombinedParsers.AnyChar — Type

AnyChar()

Parser matching exactly one Char, returning the character.

julia> AnyChar()
re"."

CombinedParsers.CharIn — Type

CharIn(x)

Parser matching exactly one character in x.

julia> a_z = CharIn('a':'z')
re"[a-z]"

julia> parse(a_z, "a")
'a': ASCII/Unicode U+0061 (category Ll: Letter, lowercase)

julia> ac = CharIn("ac")
re"[ac]"

julia> parse(ac, "c")
'c': ASCII/Unicode U+0063 (category Ll: Letter, lowercase)

CombinedParsers.CharNotIn — Type

CharNotIn(x)

Parser matching exactly one character in x.

julia> a_z = CharNotIn('a':'z')
re"[^a-z]"

julia> parse(a_z, "a")
ERROR: ArgumentError: no successfull parsing.
Stacktrace:
 [1] parse(::CharNotIn{Tuple{StepRange{Char,Int64}}}, ::String) at /home/gregor/dev/julia/CombinedParsers/src/CombinedParsers.jl:2649
 [2] top-level scope at REPL[19]:1

julia> ac = CharNotIn("ac")
re"[^ac]"

julia> parse(ac, "c")
ERROR: ArgumentError: no successfull parsing.
Stacktrace:
 [1] parse(::CharNotIn{Array{Char,1}}, ::String) at /home/gregor/dev/julia/CombinedParsers/src/CombinedParsers.jl:2649
 [2] top-level scope at REPL[19]:1

Combining Parser

CombinedParsers.Either — Type

Either(parsers...)

Parser that tries matching the provided parsers in order, accepting the first match, and fails if all parsers fail.

This parser has no custom == and hash methods because it can recurse.

julia> match(r"a|bc","bc")
RegexMatch("bc")

julia> parse(Either("a","bc"),"bc")
"bc"

julia> parse("a" | "bc","bc")
"bc"

Base.:| — Method

(|)(x::AbstractToken, y)
(|)(x, y::AbstractToken)
(|)(x::AbstractToken, y::AbstractToken)

Operator syntax for sEither(x, y).

julia> 'a' | CharIn("AB") | "bc"
|🗄... Either
├─ a
├─ [AB] CharIn
└─ bc
::Union{Char, SubString}

Base.push! — Function

Base.push!(x::Either, option)

Push option to x.options as parser tried next if x fails. Recursive parsers can be built with the push! to Either.

See also pushfirst!.

Base.pushfirst! — Function

Base.push!(x::Either, option)

Push option to x.options as parser tried first before x . Recursive parsers can be built with the push! to Either.

See also push!.

CombinedParsers.sEither — Function

sEither(x...)

Simplifying Either, flattens nested Eithers, remove Never parsers.

julia> Either('a',CharIn("AB")|"bc")
|🗄... Either
├─ a
└─ |🗄... Either
   ├─ [AB]
   └─ bc
::Union{Char, SubString}

julia> sEither('a',CharIn("AB")|"bc")
|🗄... Either
├─ a
├─ [AB]
└─ bc
::Union{Char, SubString}

See also Either

CombinedParsers.Sequence — Type

Sequence(;kw...)

Sequence keyword argument constructors transform the parsing into a named tuple.

CombinedParsers.sSequence — Function

sSequence(x...)

Simplifying Sequence, flatten Sequences, remove Always lookarounds.

julia> Sequence('a',CharIn("AB")*'b')
🗄 Sequence
├─ a
└─ 🗄 Sequence
   ├─ [AB] CharIn
   └─ b
::Tuple{Char,Tuple{Char,Char}}

julia> sSequence('a',CharIn("AB")*'b')
🗄 Sequence
├─ a
├─ [AB] CharIn
└─ b
::Tuple{Char,Char,Char}

See also Sequence

CombinedParsers.Atomic — Type

Atomic(x)

A parser matching p, and failing when required to backtrack (behaving like an atomic group in regular expressions).

Repeating

CombinedParsers.Optional — Type

Optional(parser;default=defaultvalue(result_type(parser)))

Parser that always succeeds. If parser succeeds, return result of parser with curser behind match. If parser does not succeed, return default with curser unchanged.

julia> match(r"a?","b")
RegexMatch("")

julia> parse(Optional("a", default=42),"b")
42

Base.:| — Method

(|)(x::AbstractToken{T}, default::Union{T,Missing})

Operator syntax for Optional(x, default=default).

julia> parser("abc") | "nothing"
|🗄... Either
├─ abc
└─ nothing
::SubString

CombinedParsers.Repeat — Type

Repeat(x)
Repeat(x; min=0,max=Repeat_max)
Repeat(min::Integer, x)
Repeat(min::Integer,max::Integer, x)

Parser repeating pattern x min:max times.

CombinedParsers.Lazy — Type

Lazy(x::Repeat)
Lazy(x::Optional)

Lazy x repetition matching from greedy to lazy.

julia> re"a+?"
a+?  |> Repeat |> Lazy |> regular expression combinator
::Array{Char,1}

julia> re"a??"
a??  |> Optional(default=missing) |> Lazy |> regular expression combinator
::Union{Missing, Char}

CombinedParsers.Repeat1 — Function

Repeat1(x)

Parser repeating pattern x one time or more.

Repeat1(f::Function,a...)

Abbreviation for map(f,Repeat1(a...)).

CombinedParsers.Repeat_stop — Function

Repeat_stop(p,stop)

Repeat p until stop (NegativeLookahead), not matching stop. Sets cursor before stop. Returns results of p.

julia> p = Repeat_stop(AnyChar(),'b') * AnyChar()
🗄 Sequence
├─ 🗄* Sequence |> map(IndexAt(2)) |> Repeat
│  ├─ (?!🗄) NegativeLookahead
│  │  └─ b
│  └─ . AnyChar
└─ . AnyChar
::Tuple{Array{Char,1},Char}

julia> parse(p,"acbX")
(['a', 'c'], 'b')

See also NegativeLookahead

CombinedParsers.Repeat_until — Function

Repeat\_until(p,until, with_until=false;wrap=identity)

Repeat p until stop (with Repeat_stop). and set point after stop.

Return a Vector{result_type(p)} if wrap_until==false, otherwise a Tuple{Vector{result_type(p)},result_type(until)}.

To transform the Repeat_stop(p) parser head, provide a function(::Vector{result_type(p)}) in wrap keyword argument, e.g.

julia> p = Repeat_until(AnyChar(),'b') * AnyChar()
🗄 Sequence
├─ 🗄 Sequence |> map(IndexAt(1))
│  ├─ 🗄* Sequence |> map(IndexAt(2)) |> Repeat
│  │  ├─ (?!🗄) NegativeLookahead
│  │  │  └─ b
│  │  └─ . AnyChar
│  └─ b
└─ . AnyChar
::Tuple{Array{Char,1},Char}

julia> parse(p,"acbX")
(['a', 'c'], 'X')

julia> parse(Repeat_until(AnyChar(),'b';wrap=JoinSubstring),"acbX")
"ac"

See also NegativeLookahead

Base.join — Function

Base.join(x::Repeat,delim)

Parser matching repeated x.parser separated by delim.

julia> parse(join(Repeat(AnyChar()),','),"a,b,c")
3-element Array{Char,1}:
 'a'
 'b'
 'c'

CombinedParsers.alternate — Function

optimized repeated alternations of xdelim, optionally starting/ending with delim. delim is agged as right borders. delim can be discarded in agg(result,missing,delim).

if agg is nothing, default is to aggregate delim after match is result_type(delim) <: result_type(x), if not missing.

Assertions

CombinedParsers.AtStart — Type

AtStart()

Parser succeding if and only if at index 1 with result_type AtStart.

julia> AtStart()
re"^"

CombinedParsers.AtEnd — Type

AtEnd()

Parser succeding if and only if at last index with result_type AtEnd.

julia> AtEnd()
re"$"

CombinedParsers.Always — Type

Always()

Assertion parser matching always and not consuming any input. Returns Always().

julia> Always()
re""

CombinedParsers.Never — Type

Never()

Assertion parser matching never.

julia> Never()
re"(*FAIL)"

CombinedParsers.Lookahead — Function

Lookahead(does_match::Bool, p)

PositiveLookahead if does_match==true, NegativeLookahead otherwise.

CombinedParsers.PositiveLookahead — Type

PositiveLookahead(parser)

Parser that succeeds if and only if parser succeeds, but consumes no input. The match is returned. Useful for checks like "must be followed by parser, but don't consume its match".

julia> la=PositiveLookahead("peek")
re"(?=peek)"

julia> parse(la*AnyChar(),"peek")
("peek", 'p')

CombinedParsers.NegativeLookahead — Type

NegativeLookahead(parser)

Parser that succeeds if and only if parser does not succeed, but consumes no input. parser is returned as match. Useful for checks like "must not be followed by parser, don't consume its match".

julia> la = NegativeLookahead("peek")
re"(?!peek)"

julia> parse(la*AnyChar(),"seek")
(re"(?!peek)", 's')

CombinedParsers.Lookbehind — Function

Lookbehind(does_match::Bool, p)

PositiveLookbehind if does_match==true, NegativeLookbehind otherwise.

CombinedParsers.PositiveLookbehind — Type

PositiveLookbehind(parser)

Parser that succeeds if and only if parser succeeds before cursor. Consumes no input. The match is returned. Useful for checks like "must be preceded by parser, don't consume its match".

CombinedParsers.NegativeLookbehind — Type

NegativeLookbehind(parser)

Parser that succeeds if and only if parser does not succeed before cursor. Consumes no input. nothing is returned as match. Useful for checks like "must not be preceded by parser, don't consume its match".

julia> la=NegativeLookbehind("keep")
re"(?<!peek)"

julia> parse("peek"*la,"peek")
("peek", 'p')

Printing

CombinedParsers.regex_string — Function

regex_string(x::CombinedParser)

regex_prefix(x)*regex_inner(x)*regex_suffix(x)

CombinedParsers.regex_prefix — Function

regex_prefix(x)

Prefix printed in parser tree node.

CombinedParsers.regex_inner — Function

regex_inner(x::AbstractToken)

Regex representation of x. See regex_string

CombinedParsers.regex_suffix — Function

regex_suffix(x)

Suffix printed in parser tree node.

CombinedParsers.print_constructor — Function

print_constructor(io::IO,x)

Print constructor pipeline in parser tree node.

CombinedParsers.MemoTreeChildren — Type

decurse recursive patterns