CombinedParsers.Regexp

CombinedParsers.Regexp — Module

A regular expression parser transforming a PCRE string to a CombinedParser equivalent to the regular expression.

CombinedParsers.Regexp.with_options — Function

with_options(options::AbstractString,str::AbstractString)

CombinedParsers.Regexp.@pcre_tests — Macro

@pcre_testset

Define @syntax pcre_test and @syntax pcre_tests for parsing unit test output of the PCRE library. The parser is used for testing CombinedParser and benchmarking against Regex.

Parsing Options

CombinedParsers.Regexp.ParserOptions — Type

A wrapper matching the inner parser on with_options(set_flags, unset_flags, sequence).

CombinedParsers.Regexp.FilterOptions — Type

Lazy wrapper for a sequence, masking elements in getindex with MatchingNever if any of flags are not set.

TODO: make flags a filter function?

CombinedParsers.Regexp.WithOptions — Type

A lazy element transformation type (e.g. AbstractString), getindex wraps elements in with_options(flags,...).

With parsing options

TODO: make flags a transformation function?

CombinedParsers.Regexp.OnOptionsParser — Type

Parser wrapper sequence with if_options.

Regular Expression Types

CombinedParsers.Regexp.parse_options — Function

parse_options(options::AbstractString)

Return PCRE option mask parsed from options.

Parser for flags in @re_str.

julia> CombinedParsers.Regexp.pcre_options_parser
🗄 Sequence |> map(#52)
├─ ^ AtStart
├─ 🗄+? Sequence |> map(#52) |> Repeat |> map(#44) |> Optional(default=0)
│  ├─ |🗄... Either
│  │  ├─ dupnames  |> map(Constant(0x00000040)) |> with_name(:DUPNAMES)
│  │  ├─ xx  |> map(Constant(0x01000000)) |> with_name(:EXTENDED_MORE)
│  │  ├─ i  |> map(Constant(0x00000008)) |> with_name(:CASELESS)
│  │  ├─ m  |> map(Constant(0x00000400)) |> with_name(:MULTILINE)
│  │  ├─ n  |> map(Constant(0x00002000)) |> with_name(:NO_AUTO_CAPTURE)
│  │  ├─ U  |> map(Constant(0x00040000)) |> with_name(:UNGREEDY)
│  │  ├─ J  |> map(Constant(0x00000040)) |> with_name(:DUPNAMES)
│  │  ├─ s  |> map(Constant(0x00000020)) |> with_name(:DOTALL)
│  │  ├─ x  |> map(Constant(0x00000080)) |> with_name(:EXTENDED)
│  │  ├─ B  |> map(Constant(0x00000000)) |> with_name(:BINCODE)
│  │  └─ I  |> map(Constant(0x00000000)) |> with_name(:INFO)
│  └─ ,?  |> Optional(default=missing)
└─ $ AtEnd
::UInt32

CombinedParsers.regex_escape — Function

    regex_escape(s::AbstractString)

regular expression metacharacters are escaped along with whitespace.

CombinedParsers._iterate — Method

_iterate(p::ParserWithCaptures, sequence::SequenceWithCaptures,a...)

Base.empty!(sequence) before iteration.

CombinedParsers.Regexp.ParserWithCaptures — Type

Top level parser supporting regular expression features captures, backreferences and subroutines. Collects subroutines in field subroutines::Vector and indices of named capture groups in field names::Dict.

Note

implicitly called in match

See also Backreference, Capture, Subroutine

CombinedParsers.Regexp.SequenceWithCaptures — Type

SequenceWithCaptures ensapsulates a sequence to be parsed, and parsed captures.

This struct will allow for captures a sequence-level state. For next version, a match-level state passed as _iterate argument is considered.

See also ParserWithCaptures

CombinedParsers.Regexp.Capture — Type

Capture a parser result, optionally with a name. index field is initialized when calling ParserWithCaptures on the parser.

ParserWithCaptures

CombinedParsers.Regexp.Backreference — Type

Backreference(f::Function,index::Integer)

Backreference(f::Function,name::Union{Nothing,Symbol},index::Integer)

Backreference(f::Function,name::AbstractString)

Parser matching previously captured sequence, optionally with a name. index field is recursively set when calling 'ParserWithCaptures` on the parser.

CombinedParsers.Regexp.Subroutine — Type

Parser matching preceding capture, optionally with a name. index field is recursively set when calling ParserWithCaptures on the parser.

CombinedParsers.Regexp.index — Method

index(parser::Subroutine,sequence)

Index of a subroutine. "If you make a subroutine call to a non-unique named subpattern, the one that corresponds to the first occurrence of the name is used." (what about "In the absence of duplicate numbers (see the previous section) this is the one with the lowest number."?)

CombinedParsers.Regexp.subroutine_index_reset — Method

https://www.pcre.org/original/doc/html/pcrepattern.html#SEC16

CombinedParsers.Regexp.Conditional — Type

Conditional parser, _iterate cycles conditionally on _iterate_condition through matches in field yes and no respectively.

CombinedParsers.Regexp.DupSubpatternNumbers — Type

Parser wrapper for ParserWithCaptures, setting resetindex=true in `deepmapparser(::typeof(indexedcaptures),...)`.

julia> p = re"(?|(a)|(b))\1"
🗄 Sequence |> regular expression combinator with 1 capturing groups
├─ |🗄... Either |> DupSubpatternNumbers
│  ├─ (a)  |> Capture 1
│  └─ (b)  |> Capture 1
└─ \g{1} Backreference
::Tuple{Char,AbstractString}

julia> match(p, "aa")
ParseMatch("aa", 1="a")

julia> match(p, "bb")
ParseMatch("bb", 1="b")

See also pcre doc