Transformations to any result_type

CombinedParsers.Transformation — Type

Transformation{T}(f::Function, p_) where {T}
Base.map(f::Function, Tc::Type, p::CombinedParser, a...)
Base.map(f::Function, p::CombinedParser, a...)

Parser transforming result of a wrapped parser. a... is passed as additional arguments to f (at front .

CombinedParsers.result_type — Function

result_type(x::CombinedParser)

returns the result type of a parser.

The result type is a CombinedParser type parameter. Most of the time it is type-inferred within constructors by infer_result_type.

Base.map — Function

map(f::Function, p::CombinedParser, a...)

Parser matching p, transforming parsing results (x) with function f(x,a...).

See also: get, Transformation

map(T::Type, p::CombinedParser, a...)

Parser matching p, transforming ps parsing result with constructor T(x,a...).

See also: get, Transformation

map(index::IndexAt, p::CombinedParser, a...)
map(constant, p::CombinedParser, a...)

Parser matching p, transforming ps parsing results to getindex(x,index) or constant.

See also: get, Transformation

String match results

Base.:! — Function

(!)(x::CombinedParser)

Parser Transformation getting the matched SubString.

julia> Repeat(AnyChar())
.* AnyChar |> Repeat
::Array{Char,1}

julia> !Repeat(AnyChar())
.* AnyChar |> Repeat |> !
::SubString{String}

julia> !!Repeat(AnyChar())
.* AnyChar |> Repeat |> ! |> map(intern) |> map(String)
::String

(!)(x::JoinSubstring)

Parser transformating result v -> InternedStrings.intern(v).

CombinedParsers.JoinSubstring — Type

JoinSubstring(x)
(!)(x::CombinedParser)

Parser Transformation getting the matched SubString. Transformation does not evaluate get(parser.transform,...).

CombinedParsers.MatchRange — Type

MatchRange(p::CombinedParser)

Construct a Transformation{UnitRange{Int}} resulting in p when calling get fast, Succeed iif p succeeds, if so results in sequence match index UnitRange. Transformation does not evaluate get(parser.transform,...).

CombinedParsers.Constant — Type

map_constant(constant, p::CombinedParser)
parser((p,constant)::Pair)

Construct a Transformation{<:Constant} resulting in p when calling get fast, instead of computing result from state, if parser(p) matches.

julia> parser("constant" | "fixed" => :constant)
|🗄 Either => :constant
├─ constant 
└─ fixed 
::Symbol

Note

If the Pair key is a symbol, a NamedParser is created.

julia> parser(:constant => "constant" | "fixed")
|🗄 Either |> with_name(:constant)
├─ constant 
└─ fixed 
::SubString{String}

CombinedParsers.IndexAt — Type

Struct for fast access to an index of a Transformation.

julia> using CombinedParsers.Regexp

julia> p = re"(?:a|b*)."[1]
🗄 Sequence[1]
├─ |🗄 Either
│  ├─ a 
│  └─ b*  |> Repeat
└─ [^\n] CharNotIn
::Union{Char, Array{Char,1}}

See also getindex, Sequence.

CombinedParsers.infer_result_type — Function

infer_result_type(f::Function,Tc::Type,p::CombinedParser,onerror::AbstractString,ts::Type...; throw_empty_union=true)

Used by Parser Transformations to infer result type of a parser. Throws error if type inference fails, if throwemptyunion=true.